tagger fixes
