+++ data/lexicon/adjectives/comparables.js
@@ -22,12 +22,10 @@ export default [
   'burly',
   'busy',
   'cagey',
-  'calm',
   'caring',
   'cheap',
   'cheesy',
   'chilly',
-  'clear',
   'clever',
   'cloudy',
   'cold',
@@ -48,7 +46,6 @@ export default [
   'dirty',
   'divine',
   'drunk',
-  'dull',
   'dumb',
   'early',
   'earthly',
@@ -58,8 +55,6 @@ export default [
   'essential',
   'evil',
   'extreme',
-  'fake',
-  'fancy',
   'far',
   'faulty',
   'feeble',
@@ -68,7 +63,6 @@ export default [
   'financial',
   'foamy',
   'formal',
-  'foul',
   'frail',
   'friendly',
   'full',
@@ -272,7 +266,6 @@ export default [
   'young',
   'zany',
 
-  //these ones can also be verbs, for what it's worth
   'bright',
   'broad',
   'coarse',
@@ -280,7 +273,6 @@ export default [
   'dead',
   'deaf',
   'deep',
-  'fast',
   'flat',
   'fresh',
   'great',
@@ -295,7 +287,6 @@ export default [
   'rough',
   'sad',
   'sharp',
-  'short',
   'sick',
   'smart',
   'soft',
