+++ data/lexicon/switches/adj-present.js
@@ -68,4 +68,16 @@ export default [
   'tender',
   'close',
   'last',
+  'faint',
+  'bankrupt',
+  'corrupt',
+  'fast',
+  'short',
+  'foul',
+  'fake',
+  'fancy',
+  'clear',
+  'calm',
+  'dull',
+  'stray',
 ]
