+++ data/lexicon/verbs/phrasals.js
@@ -144,6 +144,8 @@ export default [
   'chip in',
   'clean down',
   'clean up',
+  'clear up',
+  'clear out',
   'close up',
   'close off',
   'close out',
@@ -408,6 +410,7 @@ export default [
   'laugh off',
   'lay off',
   'lay on',
+  'lay down',
   'leave in',
   'leave open',
   'leave out',
