+++ src/2-two/preTagger/model/lexicon/_data.js
@@ -15,17 +15,17 @@ export default {
   "Unit|Noun": "true¦cEfDgChBinchAk9lb,m6newt5oz,p4qt,t1y0;ardEd;able1b0ea1sp;!l,sp;spo1;a,oundAt,x;on9;!b,g,i1l,m,p0;h,s;!les;!b,elvin,g,m;!es;g,z;al,b;eet,oot,t;m,up0;!s",
   "Value": "true¦a few",
   "Imperative": "true¦come here",
-  "PhrasalVerb": "true¦0:8A;1:7Z;2:8N;3:8D;4:7S;5:8P;6:7N;7:7L;8:7Y;9:88;A:8I;B:8F;C:8G;D:89;E:6S;F:6U;G:6K;H:8C;I:43;a8Wb7Ec6Dd67e66f54g4Mh41iron0j3Xk3Sl39m2Vn2To2Rp27quietEr1Rs0JtZuXvacuu6CwMyammer9zJ;ero Dip KonJ;e0k0;by,up;aQeLhKiJor0Lrit15;mp0n39pe0r7s7;eel Dip 90;aLiJn2Y;gh Jrd0;in,up;n Dr J;d2in,oF;it 6Rk7lk Krm 12sh Jt7Lv5L;aw3d2o5up;aw3in,oH;rge9sJ;e 1herG;aYeWhUiQoPrLuJypO;ckGrn J;d2in,oFup;aKiJot0y 2T;ckleEp 8L;ckEdJ;e 0R;neEp 35s55;ck LdKe Jghte68me0p o0Mre0;aw3ba4d2in,up;e 6Sy 1;by,oC;ink Jrow 76;ba4ov8up;aJe 6Qll5O;m 1r 59;ck9ke Klk J;ov8shit,u5P;aJba4d2in,oFup;ba4ft8p5Uw3;a0Pc0Oe0Ih0Ei0Bl06m05n04o03pYquar53tPuNwJ;earLiJ;ngKtch J;aw3ba4o80; by;cJit 1m 1ss0;k 5P;aPe15iOoMrKuJ;cId2U;aigh27iJ;ke 78n3R;p Jrm29;by,in,oC;nIr 1tc4A;cImp0nd Jrt 21veBy 1;ba4d2up;arIeMiLlKrJur9;ing9uc7;a3Xit 62;c7l1Dn 1;e6Lll0;ber 1rt0und like;ap 5Gow D;ash 68oke0;eep LiJow 6;cJp 1;e 5U;in,oJ;ff,v8;de17gn KngJt 62z7; al5Xle0;in,o5up;aLoJu5K;ot Jut0w 75;aw3ba4f3XoC;cIdeGk60veB;e Nll1Knd Mrv7tJ; Jtl55;d2f5Min,o5upJ;!on;aw3ba4d2in,o2Sup;oHto;al5Tout0rap5T;ilBv7;aXeUiSoOuJ;b 5Mle0n Jstl7;aLba4d2inKoJt3Qu5D;ut,v8;!to;c2MrAw3;ll Lot KuJ;g38ndB;a2Mf3Fo5;arAin,o5;ng 5Tp oJ;ff,p15;aJelBinGnt0;c61d J;o3Fup;cIt0;aYeXiVlSoQrNsyc2VuJ;ll Lt J;aJba4d2in,oFt3Xup;p42w3;ap41d2in,o5t3Vup;attle9ess KiMoJ;p 1;ah23on;iJp 66r4LurEwer 66;nt0;aKuJ;gGmp 6;ce u3My50;ck Jg0le9n 6p4Y;oFup;el 58ncilG;c4Qir 30n0ss LtKy J;ba4oC; d2c2H;aw3ba4in,oF;pJw4L;e4Kt D;arrowEerd0oJ;dBteE;aTeQiPoLuJ;ddl7lJ;l 44;cInkeyLp 6uthBve J;aJd2in,o5up;l4Aw3; wi47;ss0x 1;asur7lKss J;a1Rup;t 6;ke Kn 6rJs20x0;k 6ryB;do,fun,oHsure,up;aZeTiOoJuck0;aMc43g LoJse0;k Jse41;aft8ba4d2forw2Xin4Rov8u3U;in,oF;d 6;e 08ghtMnLsKve J;oHup;ten 57;e 1k 1; 1e3T;ave Lt KvelJ; oH;d2go,in,o5up;in,oJ;pJut;en;c7p 1sh Jtch9ughBy2B;in4Co5;eKickBnock J;d2oHup;eJyG;l 36p J;aw3ba4d2f01in,oFto,up;aLoKuJ;ic7mpG;ke3KtE;c3Tzz 1;aYeTiQoNuJ;nKrrJsh 6;y 1;kerEt J;arAd2;lJneGrse3D;d Je 1;ba4d2fast,o07up;de Kt J;ba4on,up;aw3o5;aJlp0;d Ll 2Or Jt 1;fJof;rom;in,oZu1P;cIm 1nKve Jz2J;it,to;d Jg 2VkerM;d2in,o5;k 1;aXeRive Ploss 2FoLrKunJ; f0R;ab hold,in3Wow 2Q; Jof 2E;aKb1Kit,oJrAt0Tu1F;ff,n,v8;bo5ft8hPw3;aw3bJd2in,oFrise,up,w3;a4ir2E;ar 6ek0t J;aKb1Ed2in,oJrAup;ff,n,ut,v8;cKhJl2CrAt,w3;ead;ross;d aKnJ;g 1;bo5;a0De06iXlToPrLuJ;ck Je2H;arAup;eKighten Jown9y 1;aw3oC;eJshe23; 1z7;lLol J;aJwi1W;bo5rA;d 6low 1;aKeJip0;sh0;g 6ke0mJrJttenE;e 36;gQlOnMrKsJzzle0;h 34;e Jm 1;aw3ba4up;d0isJ;h 1;e Jl 1P;aw3fOin,o5;ht ba4ure0;eOnKsJ;s 1;cLd J;fJoC;or;e D;d03l 1;cNll Jrm0t1C;ap0CbLd2in,oKtJvictim;hrough;ff,ut,v8;a4ehi2G;e J;d2oCup;a0Tdge0nd 0Xy7;aNoMrJ;aKess 6op J;aw3b01in,oH;g9w9; 0Qubl17;m 1;a06h03l00oNrJut 1E;ack9eep KoJ;ss Dwd0;by,d2in,oJup;n,ut;me MoKuntJ; o23;k 6l J;d2oC;aPbOforMin,oLtKuJ;nd8;ogeth8;n,ut,v8;th,wJ;ard;a4y;pJrAw3;art;eaKose J;in,oHup;n 6;eJipG;ck Der J;on,up;lQncel0rNsLtch KveG; in;o1Jup;h Dt J;doubt,oC;ry KvJ;e 07;aw3oF;l KmE; d2;aKba4d2oFup;ff,n,ut;rAw3;a0Qe0Il0Bo02rOuJ;bblLckl03il09lk 6ndl03rJst ZtKy 19zzB;n 0EsJ;t D;e J;ov8;anVeaTiKush J;oCup;ghPng J;aMba4d2fKin,o5uJ;nd8p;orJ;th;bo5lJrAw3;ong;teJ;n 1;k J;d2in,o5up;ch0;arQg 6iOn7oMssLttlKunce Jx D;aw3ba4;e 6; arA;k Dt 1;e 1;l 6;d2up;d 1;aOeed0oJurt0;cLw J;aw3ba4d2o5up;ck;k J;in,oH;ck0nk0stB; oOaMef 1nd J;d2ov8up;er;up;r0t J;d2in,oHup;ff,nJ;to;ck Pil0nLrgKsJ;h D;ain9e D;g Dk9; on;in,o5; o5;aw3d2oHup;ff,ut;ay;ct SdOsk LuctionB; oC;ff;arAo5;ouJ;nd;d J;d2oJup;ff,n;own;o5up;ut",
+  "PhrasalVerb": "true¦0:8B;1:8O;2:80;3:8E;4:7T;5:8Q;6:7O;7:7M;8:7Z;9:89;A:8J;B:8G;C:8H;D:8A;E:6T;F:6V;G:6L;H:8D;I:8P;a8Xb7Fc6Ed68e67f55g4Nh42iron0j3Yk3Tl39m2Vn2To2Rp27quietEr1Rs0JtZuXvacuu6DwMyammer9zJ;ero Dip KonJ;e0k0;by,up;aQeLhKiJor0Lrit15;mp0n39pe0r7s7;eel Dip I;aLiJn2Y;gh Jrd0;in,up;n Dr J;d1in,oF;it 6Sk7lk Krm 12sh Jt7Mv5M;aw3d1o5up;aw3in,oH;rge9sJ;e 2herG;aYeWhUiQoPrLuJypO;ckGrn J;d1in,oFup;aKiJot0y 2T;ckleEp I;ckEdJ;e 0R;neEp 36s56;ck LdKe Jghten 6me0p o0Mre0;aw3ba4d1in,up;e 6Ty 2;by,oC;ink Jrow 77;ba4ov8up;aJe 6Rll5P;m 2r 5A;ck9ke Klk J;ov8shit,u5Q;aJba4d1in,oFup;ba4ft8p5Vw3;a0Pc0Oe0Ih0Ei0Bl06m05n04o03pYquar54tPuNwJ;earLiJ;ngKtch J;aw3ba4o81; by;cJit 2m 2ss0;k 5Q;aPe15iOoMrKuJ;c3Dd2V;aigh27iJ;ke 79n3S;p Jrm29;by,in,oC;n38r 2tc4B;c37mp0nd Jrt 21veBy 2;ba4d1up;ar35eMiLlKrJur9;ing9uc7;a3Yit 63;c7l1Dn 2;e6Mll0;ber 2rt0und like;ap 5How D;ash 69oke0;eep LiJow 6;cJp 2;e 5V;in,oJ;ff,v8;de17gn KngJt 63z7; al5Yle0;in,o5up;aLoJu5L;ot Jut0w 76;aw3ba4f3YoC;c2MdeGk61veB;e Nll1Knd Mrv7tJ; Jtl56;d1f5Nin,o5upJ;!on;aw3ba4d1in,o2Tup;oHto;al5Uout0rap5U;ilBv7;aXeUiSoOuJ;b 5Nle0n Jstl7;aLba4d1inKoJt3Ru5E;ut,v8;!to;c2NrAw3;ll Lot KuJ;g39ndB;a2Nf3Go5;arAin,o5;ng 5Up oJ;ff,p15;aJelBinGnt0;c62d J;o3Gup;c1Yt0;aYeXiVlSoQrNsyc2WuJ;ll Lt J;aJba4d1in,oFt3Yup;p43w3;ap42d1in,o5t3Wup;attle9ess KiMoJ;p 2;ah24on;iJp 67r4MurEwer 67;nt0;aKuJ;gGmp 6;ce u3Ny51;ck Jg0le9n 6p4Z;oFup;el 59ncilG;c4Rir 31n0ss LtKy J;ba4oC; d1c2I;aw3ba4in,oF;pJw4M;e4Lt D;arrowEerd0oJ;dBteE;aTeQiPoLuJ;ddl7lJ;l 45;c16nkeyLp 6uthBve J;aJd1in,o5up;l4Bw3; wi48;ss0x 2;asur7lKss J;a1Sup;t 6;ke Kn 6rJs21x0;k 6ryB;do,fun,oHsure,up;aZeTiOoJuck0;aMc44g LoJse0;k Jse42;aft8ba4d1forw2Yin4Sov8u3V;in,oF;d 6;e 09ghtMnLsKve J;oHup;ten I;e 2k 2; 2e3U;ave Lt KvelJ; oH;d1go,in,o5up;in,oJ;pJut;en;c7p 2sh Ktch9ughBy J;d1o4W;in4Co5;eKickBnock J;d1oHup;eJyG;l 36p J;aw3ba4d1f01in,oFto,up;aLoKuJ;ic7mpG;ke3KtE;c3Tzz 2;aYeTiQoNuJ;nKrrJsh 6;y 2;kerEt J;arAd1;lJneGrse3D;d Je 2;ba4d1fast,o07up;de Kt J;ba4on,up;aw3o5;aJlp0;d Ll 2Or Jt 2;fJof;rom;in,oZu1P;cMm 2nKve Jz2J;it,to;d Jg 2VkerM;d1in,o5;k 2;aXeRive Ploss 2FoLrKunJ; f0R;ab hold,in3Wow 2Q; Jof 2E;aKb1Kit,oJrAt0Tu1F;ff,n,v8;bo5ft8hPw3;aw3bJd1in,oFrise,up,w3;a4ir2E;ar 6ek0t J;aKb1Ed1in,oJrAup;ff,n,ut,v8;cKhJl2CrAt,w3;ead;ross;d aKnJ;g 2;bo5;a0De06iXlToPrLuJ;ck Je2H;arAup;eKighten Jown9y 2;aw3oC;eJshe23; 2z7;lLol J;aJwi1W;bo5rA;d 6low 2;aKeJip0;sh0;g 6ke0mJrJttenE;e I;gQlOnMrKsJzzle0;h I;e Jm 2;aw3ba4up;d0isJ;h 2;e Jl 1P;aw3fOin,o5;ht ba4ure0;eOnKsJ;s 2;cLd J;fJoC;or;e D;d03l 2;cNll Jrm0t1C;ap0CbLd1in,oKtJvictim;hrough;ff,ut,v8;a4ehi2G;e J;d1oCup;a0Tdge0nd 0Xy7;aNoMrJ;aKess 6op J;aw3b01in,oH;g9w9; 0Qubl17;m 2;a06h03l00oNrJut 1E;ack9eep KoJ;ss Dwd0;by,d1in,oJup;n,ut;me MoKuntJ; o23;k 6l J;d1oC;aPbOforMin,oLtKuJ;nd8;ogeth8;n,ut,v8;th,wJ;ard;a4y;pJrAw3;art;eaKose J;in,oHup;n 6r I;eJipG;ck Der J;on,up;lQncel0rNsLtch KveG; in;o1Jup;h Dt J;doubt,oC;ry KvJ;e 07;aw3oF;l KmE; d1;aKba4d1oFup;ff,n,ut;rAw3;a0Qe0Il0Bo02rOuJ;bblLckl03il09lk 6ndl03rJst ZtKy IzzB;n 0EsJ;t D;e J;ov8;anVeaTiKush J;oCup;ghPng J;aMba4d1fKin,o5uJ;nd8p;orJ;th;bo5lJrAw3;ong;teJ;n 2;k J;d1in,o5up;ch0;arQg 6iOn7oMssLttlKunce Jx D;aw3ba4;e 6; arA;k Dt 2;e 2;l 6;d1up;d 2;aOeed0oJurt0;cLw J;aw3ba4d1o5up;ck;k J;in,oH;ck0nk0stB; oOaMef 2nd J;d1ov8up;er;up;r0t J;d1in,oHup;ff,nJ;to;ck Pil0nLrgKsJ;h D;ain9e D;g Dk9; on;in,o5; o5;aw3d1oHup;ff,ut;ay;ct IdOsk LuctionB; oC;ff;arAo5;ouJ;nd;d J;d1oJup;ff,n;own;o5up;ut",
   "Demonym": "true¦0:15;1:12;a0Vb0Oc0Dd0Ce08f07g04h02iYjVkTlPmLnIomHpEqatari,rCs7t5u4v3welAz2;am0Gimbabwe0;enezuel0ietnam0I;gAkrai1;aiwTex0hai,rinida0Ju2;ni0Prkmen;a5cotti4e3ingapoOlovak,oma0Spaniard,udRw2y0W;ede,iss;negal0Cr09;sh;mo0uT;o5us0Jw2;and0;a2eru0Fhilippi0Nortugu07uerto r0S;kist3lesti1na2raguay0;ma1;ani;ami00i2orweP;caragu0geri2;an,en;a3ex0Lo2;ngo0Drocc0;cedo1la2;gasy,y07;a4eb9i2;b2thua1;e0Cy0;o,t01;azakh,eny0o2uwaiI;re0;a2orda1;ma0Ap2;anO;celandic,nd4r2sraeli,ta01vo05;a2iB;ni0qi;i0oneU;aiAin2ondur0unO;di;amEe2hanai0reek,uatemal0;or2rm0;gi0;ilipino,ren8;cuadoVgyp4mira3ngli2sto1thiopi0urope0;shm0;ti;ti0;aPominUut3;a9h6o4roat3ub0ze2;ch;!i0;lom2ngol5;bi0;a6i2;le0n2;ese;lifor1m2na3;bo2eroo1;di0;angladeshi,el6o4r3ul2;gaE;azi9it;li2s1;vi0;aru2gi0;si0;fAl7merBngol0r5si0us2;sie,tr2;a2i0;li0;genti2me1;ne;ba1ge2;ri0;ni0;gh0r2;ic0;an",
   "Organization": "true¦0:4D;a3Gb2Yc2Ed26e22f1Xg1Ph1Ki1Hj1Fk1Dl18m0Wn0Jo0Gp09qu08r01sTtGuBv8w3xiaomi,y1;amaha,m13ou1w13;gov,tu2Z;a3e1orld trade organizati2S;lls fargo,st1;fie28inghou2I;l1rner br3I;gree37l street journ29m17;an halOeriz2Nisa,o1;dafo2Ol1;kswagMvo;b4kip,n2ps,s1;a tod2Yps;es3Ai1;lev33ted natio30;er,s; mobi2Qaco beQd bNeAgi frida9h3im horto2Ymz,o1witt31;shi3Xy1;ota,s r 00;e 1in lizzy;b3carpen37daily ma31guess w2holli0rolling st1Rs1w2;mashing pumpki2Tuprem0;ho;ea1lack eyed pe3Lyrds;ch bo1tl0;ys;l2n3Ds1xas instrumen1J;co,la m15;efoni0Cus;a7e4ieme2Lnp,o2pice gir5quare04ta1ubaru;rbucks,to2R;ny,undgard1;en;a2x pisto1;ls;g1Nrs;few2Ainsbury2QlesforYmsu22;.e.m.,adiohead,b6e3oyal 1yana30;b1dutch she4;ank;aders dige1Gd 1max,vl1R;bu1c1Zhot chili peppe2Nlobst2C;ll;c,s;ant30izno2I;a5bs,e3fiz28hilip morrCi2r1;emier2Audenti16;nk floyd,zza hut;psi2Btro1uge0A;br2Vchina,n2V;lant2Nn1yp12; 2ason20da2I;ld navy,pec,range juli2xf1;am;us;aAb9e6fl,h5i4o1sa,vid3wa;k2tre dame,vart1;is;ia;ke,ntendo,ss0L;l,s;c,st1Htflix,w1; 1sweek;kids on the block,york09;a,c;nd1Vs2t1;ional aca2Io,we0Q;a,cYd0O;aBcdonaldAe7i5lb,o3tv,y1;spa1;ce;b1Mnsanto,ody blu0t1;ley crue,or0O;crosoft,t1;as,subisM;dica2rcedes benz,talli1;ca;id,re;'s,s;c's milk,tt14z1Z;'ore08a3e1g,ittle caesa1K;novo,x1;is,mark; 1bour party;pres0Bz boy;atv,fc,kk,m1od1J;art;iffy lu0Moy divisi0Gpmorgan1sa;! cha07;bm,hop,n1tv;g,te1;l,rpol;asbro,ewlett pack1Ri3o1sbc,yundai;me dep1n1L;ot;tac1zbollah;hi;eneral 6hq,ithub,l5mb,o2reen d0Lu1;cci,ns n ros0;ldman sachs,o1;dye1g0E;ar;axo smith kli03encoV;electr0Km1;oto0W;a4bi,da,edex,i2leetwood mac,o1rito l0D;rd,xcX;at,nancial1restoY; tim0;cebook,nnie mae;b08sa,u3xxon1; m1m1;ob0H;!rosceptics;aiml0Be6isney,o4u1;nkin donu2po0Xran dur1;an;ts;j,w j1;on0;a,f lepp0Zll,peche mode,r spiegZstiny's chi1;ld;aIbc,hEiCloudflaBnn,o3r1;aigsli5eedence clearwater reviv1ossra06;al;ca c7inba6l4m1o0Bst06;ca2p1;aq;st;dplPg1;ate;se;ola;re;a,sco1tigroup;! systems;ev2i1;ck fil-a,na daily;r1y;on;dbury,pital o1rl's jr;ne;aEbc,eBf9l5mw,ni,o1p,rexiteeU;ei3mbardiIston 1;glo1pizza;be;ng;o2ue c1;roV;ckbuster video,omingda1;le; g1g1;oodriL;cht2e ge0rkshire hathaw1;ay;el;idu,nana republ3s1xt5y5;f,kin robbi1;ns;ic;bYcTdidSerosmith,iRlKmEnheuser-busDol,pple9r6s3utodesk,v2y1;er;is,on;hland1sociated F; o1;il;by4g2m1;co;os; compu2bee1;'s;te1;rs;ch;c,d,erican3t1;!r1;ak; ex1;pre1;ss; 5catel2ta1;ir;!-lu1;ce1;nt;jazeera,qae1;da;g,rbnb;as;/dc,a3er,tivision1;! blizz1;ard;demy of scienc0;es;ba",
   "Possessive": "true¦its,my,none,our0;!s",
-  "Noun|Verb": "true¦0:96;1:92;2:8J;3:7P;4:9D;5:9G;6:9C;7:81;8:8G;9:8W;a8Xb82c6Xd68e5Yf5Bg4Zh4Qi4Jj4Gk4Dl42m3Qn3Mo3Ip2Hques8Zr1Ss0EtYuVvQwDyCzA;ip,oA;ne,om;awn,e5Xie5Q;aJeHhEiCoAre8Y;nd0rA;k,ry,sh52;ck,mp,nApe,re,sh,tne9H;!d,g;e7LiA;p,sA;k,t2;aAed;r,th0;it,lk,rBsAt1ve,x;h,te;!ehou3ra9D;aDiCoA;iAlunte0m8te,w;ce,d;be,ew,s8;cuum,l3Z;pBr7sA;e,h0;da6gra5Wlo5S;aNeMhrLiKoIrDuBwiAy5H;n,st;nArn;e,n72;aCeBiAot,u9;bu6ck44gg0m,p;at,k,nd;ck,de,in,nsAp,v6Y;f0i82;ll,ne,p,r4Jss,t8EuA;ch,r;ck,de,e,le,me,p,re;e5Fow,u9;ar,e,ll,mp0st,xt;g,lAng2rg72s6x;k,ly;a0Lc0Fe0Ch07i05k03l01m00nZou7EpSqua6WtGuCwA;ea7KiA;ng,pe,t1;b2Dit,m,ppBrAspe5;ge,pri3v2L;lAo67;e64y;aJeIiHoErBuAy2;dy,ff,mb2;a7HeBiAo59ugg2;ke,ng;am,ss,t1;ckBp,rA;e,m;ing,pi2;ck,nk,t1;m,p;ck,in,ke,ll,mp,nd,p2rAte,y;!e,t;aFeed,iElDoCrAur,y;ay,e4RinAu4;g,k2;ns7Nt;a6Fit;ll,n,r7Jte;n,rk;ap,ee1Pow;a6Be44i2o4R;eep,iAou1;ce,p,t;ateboa6ZiA;!p;de,gnAp,ze;!al;aDeCiBoAuff2;ck,p,re,t,w;ft,p,v0;d,i3N;ck,de,pe,re,ve;aBed,nArv1Gt;se,t2F;l,r1t;aEhedu2oDrA;aBeAo3O;en,w;pe,t1;re,ut;le,n,r0N;cAil;k,rifi4;aWeEiDoBuA;b,in,le,n,s5F;a9ck,ll,oApe,u6;f,t;de,ng,p,s1O;aQcPdo,el,fNgMje5lLmJnIo13pGque9sCturn,vAwa6A;eAi1Y;al,r3;er6IoCpe5tBuA;lt,me;!a4O;l6Frt;air,eaAly,o4M;l,t;dezvo2Ot;aAedy;ke,rk;ea3i57;a5Xist0r54;act6Cer1Morm,uA;nd,se;a2Wo5Uru8;ch,d,s5V;c15ge,i3ke,l4YnApXtX;ge,k;a03eZhXiUlQoKrBsy1uA;mp,n1rcha3sh;aHeFiEoAu49;be,ceCdu4fi2grAje5mi3p,te9;amAe6C;!me;ed,ss;ce,de,nt;sAy;er5Ss;cti4i3;iElCoBp,re,sAw0;e,i5Ft;l,p;iAl;ce,sh;nt,s5C;aBce,e2RuA;g,mp,n7;ce,nAy;!t;ck,lBnApe,t,vot;!e;e,ot;a3oA;ne,tograph;ak,eCnBrAt;fu4Lm8;!c2P;!l,r;ckFiEnDrBsAt1u3;s,te;e5PtA;!y;!ic;nt,r;!a7;bCff0il,oBrAutli3A;b8d0ie5J;ze;je5;a47eCoA;d,tA;e,i4;ed,gle5rd,t;aHeFiEoBuA;rd0;ck,d3Bld,nAp,uth,ve;it4UkA;ey;lk,n3Urr4Sss,x;asu3En4CrAss;ge,it;il,nCp,rk3GsBtA;ch,t0;h,k;da6n0oeuv39;aHeFiDoAump;aBbby,ck,g,oAve;k,t;d,n;cAft,m8nEst;en3k;aAc0Ie1vG;ch,d,k,p,se;bBcAnd,p,t1un1;e,k;el,o2I;eBiAno3Y;ck,ll,ss;el,y;aBo1GuA;d7i4mp;m,zz;mpFnBr3QssA;ue;cr1Idex,fluCha2k,se25terAvoi4;e9fa4viA;ew;en4;a5le1Z;aFeDiCoBuA;g,nt;l2Xno22ok,p,r3st,u3;ghlight,ke,re,t;aAd7lp;d,t;ck,m,ndBrAte;bo3Jm,ne40ve9;!le;aKeek,loJoHrCuA;arAe3Xi0Tn;antee,d;aCiBoAumb2;om,u2P;nd,p;dAsp;e,ua6;of,ssA;ip;ss,w;in,me,ng,s,te,ze;aVeSiNlKoGrCuA;ck,el,nAss,zz;c2Vd;aBoAy;st,wn;cAgme,me;tu1T;cCg,il,ld,ol,rA;ce,e1YmA;!at;us;aBe1Uip,oAy;at,ck,od,w;g,ke,me,re,sh,vo16;eDgClBnAre,sh,t,x;an4i1Q;e,m,t0;ht,u1I;ld;aBeAn4;d,l;r,tu1E;ce,il,ll,rm,vo2J;cho,d7nIsGxCyeA;!baA;ll;cCerci3hib8pAtra5;eriAo0O;en4me2Z;el,han7;caAtima6;pe;count0d,gine0vy;aWePiJoGrBuAye;b,el,mp,pli2K;aDeCiBoA;ne,p;ft,nk,p,ve;am,ss;ft,g,in;cBd7ubt,wnloA;ad;k,u07;p,sCt1vA;e,iAor4;de;char7h,liBpA;at1lay,u6;ke;al,ba6cEfeDlCma14pos8siBtaA;il;gn,re;ay,ega6;at,ct;liZrA;ea3;b,ma7n4rAte;e,t;a09ent08h01irc2lYoGrCuA;be,e,rAt;e,l,ve;aCeBoAu0Gy;p,ss,wd;d8ep;ck,ft,sh;a1de,in,lPmInCok,py,re,st,uAv0;gh,nAp2;sXt;ceDdu5fli5glomeEsCtAveJ;a5rA;a9ol;e1Stru5;ntArn;ra6;bEfoDmCpA;leAou0Eromi3;me1N;a0Ce1Mit,u6;rt;at,iA;ne;lap3oA;r,ur;aBiA;ck,p;im,w;aBeAip;at,ck,er;iDllen7mpi0InBrAuffe0P;ge,m,t;ge,nA;el;n,r;er,re;ke,ll,mp,noe,pCrSsBt1u3ve;se;h,t;!tuA;re;aYeUiSlPoLrDuAypa0V;bb2ck2dgBff0lAmp,rn,st,zz;ly;et;anGeEiCoadBuA;sh;ca9;be,d7;ge;aAed;ch,k;ch,d;aCg,mb,nBoAss,tt2x,ycott;k,st,t;d,e;rd,st;aBeAitz,oTur;nd;me;as,d,ke,nd,tA;ch,e;aCef,nAt;d,ef8;it;r,t;ck,il,lan4nErgCsBtt2;le;e,h;aAe;in;!d,g,k;c02dUffilTge,iSlt0nPppMrIsFttDucBwaA;rd;tiA;on;aAempt;ck;k,sA;i9ocM;st;chBmoA;ur;!iA;ve;eBroa1;ch;al;chAg0sw0;or;er;d,m,r;ia6;dEvA;an4oA;ca6;te;ce;i5reA;ss;ct;cAhe,t;eCoA;rd,uA;nt;nt,ss",
+  "Noun|Verb": "true¦0:96;1:92;2:8J;3:7P;4:9D;5:9G;6:9C;7:81;8:8G;9:8W;a8Xb82c6Xd68e5Yf5Bg4Zh4Qi4Jj4Gk4Dl42m3Qn3Mo3Ip2Hques8Zr1Ss0EtYuVvQwDyCzA;ip,oA;ne,om;awn,e5Xie5Q;aJeHhEiCoAre8Y;nd0rA;k,ry,sh52th;ck,mp,nApe,re,sh,tne9H;!d,g;e7LiA;p,sA;k,t2;aAed;r,th0;it,lk,rBsAt1ve,x;h,te;!ehou3ra9D;aDiCoA;iAlunte0m8te,w;ce,d;be,ew,s8;cuum,l3Z;pBr7sA;e,h0;da6gra5Wlo5S;aNeMhrLiKoIrDuBwiAy5H;n,st;nArn;e,n72;aCeBiAot,u9;bu6ck44gg0m,p;at,k,nd;ck,de,in,nsAp,v6Y;f0i82;ll,ne,p,r4Jss,t8EuA;ch,r;ck,de,e,le,me,p,re;e5Fow,u9;ar,e,ll,mp0st,xt;g,lAng2rg72s6x;k,ly;a0Lc0Fe0Ch07i05k03l01m00nZou7EpSqua6WtGuCwA;ea7KiA;ng,pe,t1;b2Dit,m,ppBrAspe5;ge,pri3v2L;lAo67;e64y;aJeIiHoErBuAy2;dy,ff,mb2;a7HeBiAo59ugg2;ke,ng;am,ss,t1;ckBop,p,rA;e,m;ing,pi2;ck,nk,t1;m,p;ck,in,ke,ll,mp,nd,p2rAte,y;!e,t;aFeed,iElDoCrAur,y;ay,e4RinAu4;g,k2;ns7Nt;a6Fit;ll,n,r7Jte;n,rk;ap,ee1Pow;a6Be44i2o4R;eep,iAou1;ce,p,t;ateboa6ZiA;!p;de,gnAp,ze;!al;aDeCiBoAuff2;ck,p,re,t,w;ft,p,v0;d,i3N;ck,de,pe,re,ve;aBed,nArv1Gt;se,t2F;l,r1t;aEhedu2oDrA;aBeAo3O;en,w;pe,t1;re,ut;le,n,r0N;cAil;k,rifi4;aWeEiDoBuA;b,in,le,n,s5F;a9ck,ll,oApe,u6;f,t;de,ng,p,s1O;aQcPdo,el,fNgMje5lLmJnIo13pGque9sCturn,vAwa6A;eAi1Y;al,r3;er6IoCpe5tBuA;lt,me;!a4O;l6Frt;air,eaAly,o4M;l,t;dezvo2Ot;aAedy;ke,rk;ea3i57;a5Xist0r54;act6Cer1Morm,uA;nd,se;a2Wo5Uru8;ch,d,s5V;c15ge,i3ke,l4YnApXtX;ge,k;a03eZhXiUlQoKrBsy1uA;mp,n1rcha3sh;aHeFiEoAu49;be,ceCdu4fi2grAje5mi3p,te9;amAe6C;!me;ed,ss;ce,de,nt;sAy;er5Ss;cti4i3;iElCoBp,re,sAw0;e,i5Ft;l,p;iAl;ce,sh;nt,s5C;aBce,e2RuA;g,mp,n7;ce,nAy;!t;ck,lBnApe,t,vot;!e;e,ot;a3oA;ne,tograph;ak,eCnBrAt;fu4Lm8;!c2P;!l,r;ckFiEnDrBsAt1u3;s,te;e5PtA;!y;!ic;nt,r;!a7;bCff0il,oBrAutli3A;b8d0ie5J;ze;je5;a47eCoA;d,tA;e,i4;ed,gle5rd,t;aHeFiEoBuA;rd0;ck,d3Bld,nAp,uth,ve;it4UkA;ey;lk,n3Urr4Sss,x;asu3En4CrAss;ge,it;il,nCp,rk3GsBtA;ch,t0;h,k;da6n0oeuv39;aHeFiDoAump;aBbby,ck,g,oAve;k,t;d,n;cAft,m8nEst;en3k;aAc0Ie1vG;ch,d,k,p,se;bBcAnd,p,t1un1;e,k;el,o2I;eBiAno3Y;ck,ll,ss;el,y;aBo1GuA;d7i4mp;m,zz;mpFnBr3QssA;ue;cr1Idex,fluCha2k,se25terAvoi4;e9fa4viA;ew;en4;a5le1Z;aFeDiCoBuA;g,nt;l2Xno22ok,p,r3st,u3;ghlight,ke,re,t;aAd7lp;d,t;ck,m,ndBrAte;bo3Jm,ne40ve9;!le;aKeek,loJoHrCuA;arAe3Xi0Tn;antee,d;aCiBoAumb2;om,u2P;nd,p;dAsp;e,ua6;of,ssA;ip;ss,w;in,me,ng,s,te,ze;aVeSiNlKoGrCuA;ck,el,nAss,zz;c2Vd;aBoAy;st,wn;cAgme,me;tu1T;cCg,il,ld,ol,rA;ce,e1YmA;!at;us;aBe1Uip,oAy;at,ck,od,w;g,ke,me,re,sh,vo16;eDgClBnAre,sh,t,x;an4i1Q;e,m,t0;ht,u1I;ld;aBeAn4;d,l;r,tu1E;ce,il,ll,rm,vo2J;cho,d7nIsGxCyeA;!baA;ll;cCerci3hib8pAtra5;eriAo0O;en4me2Z;el,han7;caAtima6;pe;count0d,gine0vy;aWePiJoGrBuAye;b,el,mp,pli2K;aDeCiBoA;ne,p;ft,nk,p,ve;am,ss;ft,g,in;cBd7ubt,wnloA;ad;k,u07;p,sCt1vA;e,iAor4;de;char7h,liBpA;at1lay,u6;ke;al,ba6cEfeDlCma14pos8siBtaA;il;gn,re;ay,ega6;at,ct;liZrA;ea3;b,ma7n4rAte;e,t;a09ent08h01irc2lYoGrCuA;be,e,rAt;e,l,ve;aCeBoAu0Gy;p,ss,wd;d8ep;ck,ft,sh;a1de,in,lPmInCok,py,re,st,uAv0;gh,nAp2;sXt;ceDdu5fli5glomeEsCtAveJ;a5rA;a9ol;e1Stru5;ntArn;ra6;bEfoDmCpA;leAou0Eromi3;me1N;a0Ce1Mit,u6;rt;at,iA;ne;lap3oA;r,ur;aBiA;ck,p;im,w;aBeAip;at,ck,er;iDllen7mpi0InBrAuffe0P;ge,m,t;ge,nA;el;n,r;er,re;ke,ll,mp,noe,pCrSsBt1u3ve;se;h,t;!tuA;re;aYeUiSlPoLrDuAypa0V;bb2ck2dgBff0lAmp,rn,st,zz;ly;et;anGeEiCoadBuA;sh;ca9;be,d7;ge;aAed;ch,k;ch,d;aCg,mb,nBoAss,tt2x,ycott;k,st,t;d,e;rd,st;aBeAitz,oTur;nd;me;as,d,ke,nd,tA;ch,e;aCef,nAt;d,ef8;it;r,t;ck,il,lan4nErgCsBtt2;le;e,h;aAe;in;!d,g,k;c02dUffilTge,iSlt0nPppMrIsFttDucBwaA;rd;tiA;on;aAempt;ck;k,sA;i9ocM;st;chBmoA;ur;!iA;ve;eBroa1;ch;al;chAg0sw0;or;er;d,m,r;ia6;dEvA;an4oA;ca6;te;ce;i5reA;ss;ct;cAe,he,t;eCoA;rd,uA;nt;nt,ss",
   "Actor": "true¦0:1C;a18b11c0Od0Hex0Gf0Bg08h07i03journ02lZmTnurse,oRpIquGrAs4t3underwrDv2w1;ork0rC;eterinari0QiceD;aoiseach,echn0Hherapi1AranscriB;ail1Gcienti19e4ing0ol11p2u1;lt0Nperv1Arge2;eciWokespers1;on;cretary,rgeant1;! maj1B;abbi,e1oof0unning 7;ar4c1port0sear0Rverend;e2ru1;it0;ptioni0Z; admir12;arter1een;back;ast12h8l6oli5r1sychologi0V;actition0eside13i2o1;duc0fess10gramm0speZ;me minBnce1;!ss;c00tX;a1umb0;nn0y0;otograph0ysU;ffic0p1rganiz0;er0Rtometri0L;a5e4i1usR;n2s1;sus,t0;ist0;chanic,rcha0P;gistrate,k0nag0y0M;a2eYieutenant1ord;! gener0H;bor0dy,wy0;ali0B;llu0Fn1;s1vestig0F;peEt1;all0ruD;airdress0ousek3;arden0oal1;ie,k1teL;eep0;arm0i1;eld marsh05g3r1tt0;e1st lady;fig1mE;ht0;amin0cellency;aUe4i1o2riv0utcheQ;et2re1;ct01;ici8;al0fenc2put1sign0tective,velop0;ee,y;em5;aBeo,fo,h8le7o1;lonel,m4n1ordinUrporSunci9wboy;gress1stable;m1wom1;an;ma1pos0;nd0;an0rk;ance2eerle1;ad0;llM;ptain,r1;etak0pent0;aron5r2ut1;ch0;i1ok0;cklay0ga1;di0;!e1;ss;ccounDd6geEn2r1ssisDttendaEyatullah;chitect,ti4;aly3im9nou1;nc0;er;st;mi2v1;is5;ni2r1;al;str1;at1;or;ta1;nt",
-  "Adj|Noun": "true¦a0Wb0Rc0Gde0Fe0Af04g03h02i00juZlXmRnPoOpKrHsAt6u3va1w0;atershed,elcome;gabo3nilla,ria0;b0Int;ndergr0pstairs;adua0Oou0;nd;a2e0oken,ri0Q;en,r0;min0Oror0G;boo,n;e4ist03o2quat,ta1u0well;bordina0Hper4;b08ndard;cial0Bl0vereign;e,ve0M;cret,n0ri0H;ior;ac07e0outiNubbiG;ar,b7la00nt0Ep0;resentaZublican;a2eriodic0Cotenti0Cr0;emi0incip0Bo;er,um;rall2tie0D;ffi01pposi04v08;aggRov0;el;aTe3in2o0;biSdernWr0;al,tal;iature,or;di0tr06;an,um;attGiberZu0;sh;stice,veniL;deWmpressionOn0;cumbeZdividuVnoYstaZ;omel9umdrum;enious,old;a3e1i0luid;ne;llow,m0;aDinH;ir,t,vo0;riJuriJ;l2pRx0;c0ecu7pM;ess;d0iF;er;mographMriva3;hiDlassLo1rude,unn0;ing;m3n1opera0;tive;cre8stitueGtemporary,vertab0;le;m1p0;anion,lex;er1un0;ist;ci6;lank,o3r0;i1u0;te;ef;ttom,urgeois;cadem5d2l1nim0rab;al;ert;oles0ult;ce0;nt;ic",
+  "Adj|Noun": "true¦a0Wb0Rc0Gde0Fe0Af04g03h02i00juZlXmRnPoOpKrHsAt6u3va1w0;atershed,elcome;gabo3nilla,ria0;b0Int;ndergr0pstairs;adua0Oou0;nd;a2e0oken,ri0Q;en,r0;min0Oror0G;boo,n;e4ist03o2quat,ta1u0well;bordina0Hper4;b08ndard;cial0Bl0vereign;e,ve0M;cret,n0ri0H;ior;ac07e0outiNubbiG;ar,b7la00nt0Ep0;resentaZublican;a2eriodic0Cotenti0Cr0;emi0incip0Bo;er,um;rall2tie0D;ffi01pposi04v08;aggRov0;el;aTe3in2o0;biSdernWr0;al,tal;iature,or;di0tr06;an,um;attGiberZu0;sh;stice,veniL;deWmpressionOn0;cumbeZdividuVnoYstaZ;alf,omel9umdrum;enious,old;a3e1i0luid;ne;llow,m0;aDinH;ir,t,vo0;riJuriJ;l2pRx0;c0ecu7pM;ess;d0iF;er;mographMriva3;hiDlassLo1rude,unn0;ing;m3n1opera0;tive;cre8stitueGtemporary,vertab0;le;m1p0;anion,lex;er1un0;ist;ci6;lank,o3r0;i1u0;te;ef;ttom,urgeois;cadem5d2l1nim0rab;al;ert;oles0ult;ce0;nt;ic",
   "Person|Noun": "true¦a07b01cYdRePfOgMhJjFkiElDmBolive,p7r4s3trini00v1wa0;ng,rd;an,enus,iol0;a,et;ky,on5umm02;ay,e1o0uby;bin,d,se;ed,x;a2e0ol;aHn0;ny;ge,tM;a0eloR;x,ya;a9eo,iE;ng,tL;a2e1o0;lDy;an,w3;de,smi4y;a0iKol8;ll,z0;el;ail,e0;ne;aith,ern,lo;a0dDmir,ula,ve;rl;a4e3i1ol0;ly;ck,x0;ie;an,ja;i0wn;sy;h0liff,rystal;ari0in,ristian;ty;ak4e3i2r0;an0ook;dy;ll;nedict,rg;er;l0rt;fredo,ma",
   "Person|Verb": "true¦b8ch7dr6foster,gra5hope,ja9lan4ma2ni9ollie,p1rob,s0tra4wade;pike,t5ue;at,eg,ier2;ck,r0;k,shal;ce;ce,nt;ew;ase,u1;iff,l1ob,u0;ck;aze,ossom",
   "Pronoun": "true¦'em,elle,h3i2me,she4th0us,we,you;e0ou;m,y;!l,t;e0im;!'s",
-  "Singular": "true¦0:4L;1:5D;2:5C;3:4R;4:4Z;5:4W;6:4V;a4Tb48c3Ad2We2Pf2Fg24h1Rin1Njel4k1Ml1Im18n15o11p0Jqu0Ir09sRtHuEvAw7x 53;a7ha3Com2B;f0i4Xt0Ay7;! arou4G;arn4He8o7;cabula42l55;gIr7;di6t1H;nc35p2Sr8s 7;do3Ts58;bani2in1; rex,aFeEhing40iCoBr9u8v7;! show;m2In3rntGto13;agedy,ib7o46;e,u2P;p3rq3E;c,er,m7;etB;am,mp3A;ct3le5x return;aNcLeKhor4Pi2kIoHtCu9weethe1Hy7;ll7n27st4S;ab2Q;b8nri19per bowl,r7;f0roga2;st3Ftot1;aAep9ipe3So1Ara8udent7;! lo1K;ng0te4F;fa1Fmo1F;ff0t31;ft ser4Hloi43meo15;elet12i7;er,ll,rm3N;ab0Ucurity gu2F;e5ho7;l32ol;la35ndwi0I;av0WeAhetor3i8o7;de5om;te,v7;erb0N;bAc9f7publ3r0Ospi2;er7orm0;e5r1;it1ord label;a2u45;estion mark,ot2A;aLeJhIiFlDort1r8u7yram1D;ddi3Eppy,rpo0J;eAie3Ko7;bl3Ys7;pe6t7;a2itu2;diction,mi0Eroga3Tss relea0E;a7ebisci2;q29te;cn3e8g7;!gy;!r;armaci3AotocoF;anut,dest1r7t1;cen3Jsp3K;n8r2Tte7;!nt;el2Top4;bj3Gc8pia2rde1thers,ve7wn0;n,rview;cu7e0F;pi0;a8it26ot7umb0;a27hi2S;n2Brra3B;aDeCiBo8u7é0G;m0Sr1;m8nopo4pPrni2Osq1Qt7;h0i38;!my;li0Xn09;d3nu,t1;mm1n8te7yf4;ri1;!d11;aurea2i9os0u7;ddi2n7;ch;ght bulb,p0B;eyno2ittJ;c9dices,itia2Vni2Bse6te8vert7;eb1N;l0Mnt;en2Side5;aIeaDighBo7uman right,ygie10;me8sp1Ktb7;ed;! r7;un; scho12ri7;se;d8v7;en; st8pho7;ne;art;ndful,ze;aFeDirl1KlaOoCr8u7;l4y;an7enadi0id;a16d7; slam,fa7mo7;th0;d,lf0;lat0Cntlem7;an;df4r7;l3n1D;aFeEiClBol4r8un7;er1;ee market,i8on7;ti0;e16ga2;ow0u2;nan7ref4;ci0;lla,t14;br3mi4n0Uth0;conoCffBgg,lecto0MnAs1Yth3ven9x8yel7;id;ampSempl0Ite5;i16t;er1L;e6i1K;my;aIeGiCo9r8u7;de,mp truck;agonf4i0;c8g1Fi4or,ssi0wn7;si0L;to0Aumen9;a9gni8nn0s7vide0N;conte5incen1Ftri6;ta09;le6;ath,c1f7ni1terre5;ault 04err1;d7rX;!dy;aWeThMiLlJoBr7;edit c9it3uc7;ib7;le;ard;efficDke,lBmmuniqLn9pi0rr1t11u7yo2;ri0s7;in;ne6s7;ervatoUuG;ic,lPum7;ni0K;ie5;er7ie5;gy,ic;ty,vil wK;aBeqAoco9r7;istmas car7ysanthemum;ol;la2;ue;ndeli0racter7;ist3;i8llCr7;e1tifica2;liZ;hi0naDpCrAshi0t7ucus;erpi7hedr1;ll7;ar;!bohyd7ri0;ra2;it1;ry;aNeMiKlemJoFrBu7;ddhiWn9r7tterf4;glar7i1;!y;ny;eak9i8o7;!th0;de;faPthroA;dy,ro9wl,y7;!frie7;nd;ugh;ish;cyc7oF;liI;an,l4;nki9r7;!ri0;er;ng;cSdMllKnHppeti2rGsDt9u7;nt,to7;psy;h8t3;ic;ie7le2;st;ce5pe6;ct;nt;ray;ec8oma4ti8;ly;do2;er7y;gy; hominBj8van7;tage;ec7;ti7;ve;em;cru1e8qui7;tt1;ta2;te;al",
+  "Singular": "true¦0:4K;1:5C;2:5B;3:4Q;4:4Y;5:4V;6:4U;a4Sb47c3Ad2We2Pf2Fg24h1Rin1Njel4k1Ml1Im18n15o11p0Jqu0Ir09sRtHuEvAw7x 52;a7ha3Com2B;f0i4Wt0Ay7;! arou4F;arn4Ge8o7;cabula41l54;gIr7;di6t1H;nc35p2Sr8s 7;do3Ss57;bani2in1; rex,aFeEhing3ZiCoBr9u8v7;! show;m2In3rntGto13;agedy,ib7o45;e,u2P;p3rq3D;c,er,m7;etB;am,mp39;ct3le5x return;aNcLeKhor4Oi2kIoHtCu9weethe1Hy7;ll7n27st4R;ab2Q;b8nri19per bowl,r7;f0roga2;st3Etot1;aAep9ipe3Ro1Ara8udent7;! lo1K;ng0te4E;fa1Fmo1F;ff0t30;ft ser4Gloi42meo15;elet12i7;er,ll,rm3M;ab0Ucurity gu2F;e5ho7;l31ol;la34ndwi0I;av0WeAhetor3i8o7;de5om;te,v7;erb0N;bAc9f7publ3r0Ospi2;er7orm0;e5r1;it1ord label;a2u44;estion mark,ot29;aLeJhIiFlDort1r8u7yram1D;ddi3Dppy,rpo0J;eAie3Jo7;bl3Xs7;pe6t7;a2itu2;diction,mi0Eroga3Sss relea0E;a7ebisci2;q28te;cn3e8g7;!gy;!r;armaci39otocoF;anut,dest1r7t1;cen3Isp3J;n8r2Ste7;!nt;el2Sop4;bj3Fc8pia2rde1thers,ve7wn0;n,rview;cu7e0F;pi0;a8it25ot7umb0;a26hi2R;n2Arra3A;aDeCiBo8u7é0G;m0Sr1;m8nopo4pPrni2Nsq1Pt7;h0i37;!my;li0Xn09;d3nu,t1;mm1n8te7yf4;ri1;!d11;aurea2i9os0u7;ddi2n7;ch;ght bulb,p0B;eyno2ittJ;c9dices,itia2Uni2Ase6te8vert7;eb1M;l0Mnt;en2Ride5;aIeaDighBo7uman right,ygie0Z;me8sp1Jtb7;ed;! r7;un; scho11ri7;se;d8v7;en; st8pho7;ne;art;ndful,ze;aFeDirl1JlaOoCr8u7;l4y;an7enadi0id;a15d7; slam,fa7mo7;th0;d,lf0;lat0Cntlem7;an;df4r7;l3n1C;aFeEiClBol4r8un7;er1;ee market,i8on7;ti0;e15ga2;ow0u2;nan7ref4;ci0;lla,t13;br3mi4n0Tth0;conoCffBgg,lecto0LnAs1Xth3ven9x8yel7;id;ampSempl0Hte5;i15t;er1K;e6i1J;my;aIeGiCo9r8u7;de,mp truck;agonf4i0;c8g1Ei4or,ssi0wn7;si0K;to09umen9;a9gni8nn0s7vide0M;conte5incen1Etri6;ta08;le6;ath,c1f7ni1terre5;ault 03err1;d7rW;!dy;aVeShLiKlIoBr7;edit c9it3uc7;ib7;le;ard;efficCke,lAmmuniqKns9pi0rr1t10u7yo2;ri0s7;in;ervatoUuG;ic,lPum7;ni0K;ie5;er7ie5;gy,ic;ty,vil wK;aBeqAoco9r7;istmas car7ysanthemum;ol;la2;ue;ndeli0racter7;ist3;i8llCr7;e1tifica2;liZ;hi0naDpCrAshi0t7ucus;erpi7hedr1;ll7;ar;!bohyd7ri0;ra2;it1;ry;aNeMiKlemJoFrBu7;ddhiWn9r7tterf4;glar7i1;!y;ny;eak9i8o7;!th0;de;faPthroA;dy,ro9wl,y7;!frie7;nd;ugh;ish;cyc7oF;liI;an,l4;nki9r7;!ri0;er;ng;cSdMllKnHppeti2rGsDt9u7;nt,to7;psy;h8t3;ic;ie7le2;st;ce5pe6;ct;nt;ray;ec8oma4ti8;ly;do2;er7y;gy; hominBj8van7;tage;ec7;ti7;ve;em;cru1e8qui7;tt1;ta2;te;al",
   "Preposition": "true¦-,aNbKcJdIexcept,fGinFmid,notwithstandiToDpUqua,sCt7u4v2w0;/o,herePith0;! whFin,oT;ersus,i0;a,s-a-vis;n1p0;!on;like,til;h1ill,oward0;!s;an,r0;ough0u;!oJ;ans,ince,o that,uch D;f0n1ut;!f;!to;or,r0;om;espite,own,u3;hez,irca;ar1e0oAy;sides,tween;ri6;bo7cross,ft6lo5m3propos,round,s1t0;!op;! long 0;as;id0ong0;!st;ng;er;ut",
   "Uncountable": "true¦0:25;1:2D;a2Kb2Cc20d1Re1Hf1Bg14h0Xi0Sjewel10k0Rl0Mm0Cn0Ao09p00rVsIt8v6w2you guys;a4i3oo2;d,l;ldlife,ne;rm7t1;ernacul1Zi2;neg1Yol1Etae;eAh8ime off,oothpas1Qr3un2yranny;a,gst20;aff2Dea1Ao3ue nor2;th;o03u2;ble2se1It;!shoot2C;er2und1;e,mod2J;a,nnis;aCcBeAh9il8ki7o6p5t3u2weepstak0;g1Mnshi28shi;ati05e2;am,el;ace2Beci0;ap,cc1meth23;n,ttl0;k,v1;eep,ingl0or13;na17ri0;ene09isso13;d0Nfe22l2nd,t0B;m1Nt;a5e3ic2;e,ke0X;c2ins,laxa0Usearch;ogni0Trea0T;bi0in;a8e7hys6last1No4re2;amble,mis0s2ten1R;en1Qsu0C;l2rk;i1ZyE; 0Zi1Z;a1Vtr05;nt2ti0E;i0s;bstetri1Vvercrowd1Hxyg10;a2ews;il polTtional securi1K;a9e7o4u2;m2s1A;ps;n2o1D;ey,o2;gamy;a2cha08rchandi1Atallurgy;sl0t;chine2thema1J; learn15ry;aught1e5i4ogi3u2;ck,g0W;c,s1F;ce,ghtn11nguis1EteraOv1;ath1isuOss;ara07indergart0Inowled0T;ce,gnor5mp4n2;formaYter2;net,sta04;atiQort2rov;an15;a5eHisto4o2ung1;ckey,mework,ne2rserad6spitali0X;s0Wy;ry;ir,libZppiFs2;h2te;ish;ene5l4o3r2um,ymnas0Z;aCeZ;lf,re;ut02yce0F; 2ti0W;edit0Hpo2;ol;aNicGlour,o4urni2;tu2;re;od,rgive2uri1wl;ne2;ss;cono0JducaAlectr8n6quip7thi0Nvery5x2;ist3per2;ti09;en0H;body,o06th05;joy2tertain2;ment;ici07o2;ni0F;tiR;e8i5o3raugh2ynas04;ts;pe,wnstai2;rs;abet0s2;honYrepu2;te;b2miU;ut;aCelciBh8ivi04l6o2urrency;al,ld w4n2ral,ttGuscoA;fusiFt 2;ed;ar;assiZoth0;es;aos,e3ick2;en;eJwH;us;d,rL;a6i4lood,read,u2;ntDtt1;er;lliarHs2;on;g2ss;ga2;ge;cIdviHeroEirDm9n4ppeal court,rithmet3spi2thleJ;rin;ic;i5y2;o3th2;ing;ne;se;en4n2;es2;ty;ds;craft;bi7d2nau6;yna2;mi5;ce;id,ous2;ti2;cs",
   "SportsTeam": "true¦0:1A;1:1H;2:1G;a1Eb16c0Td0Kfc dallas,g0Ihouston 0Hindiana0Gjacksonville jagua0k0El0Bm01newToQpJqueens parkIreal salt lake,sAt5utah jazz,vancouver whitecaps,w3yW;ashington 3est ham0Rh10;natio1Oredski2wizar0W;ampa bay 6e5o3;ronto 3ttenham hotspur;blue ja0Mrapto0;nnessee tita2xasC;buccanee0ra0K;a7eattle 5heffield0Kporting kansas0Wt3;. louis 3oke0V;c1Frams;marine0s3;eah15ounG;cramento Rn 3;antonio spu0diego 3francisco gJjose earthquak1;char08paA; ran07;a8h5ittsburgh 4ortland t3;imbe0rail blaze0;pirat1steele0;il3oenix su2;adelphia 3li1;eagl1philNunE;dr1;akland 3klahoma city thunder,rlando magic;athle0Mrai3;de0; 3castle01;england 7orleans 6york 3;city fc,g4je0FknXme0Fred bul0Yy3;anke1;ian0D;pelica2sain0C;patrio0Brevolut3;ion;anchester Be9i3ontreal impact;ami 7lwaukee b6nnesota 3;t4u0Fvi3;kings;imberwolv1wi2;rewe0uc0K;dolphi2heat,marli2;mphis grizz3ts;li1;cXu08;a4eicesterVos angeles 3;clippe0dodDla9; galaxy,ke0;ansas city 3nE;chiefs,roya0E; pace0polis colU;astr06dynamo,rockeTtexa2;olden state warrio0reen bay pac3;ke0;.c.Aallas 7e3i05od5;nver 5troit 3;lio2pisto2ti3;ge0;broncZnuggeM;cowbo4maver3;ic00;ys; uQ;arCelKh8incinnati 6leveland 5ol3;orado r3umbus crew sc;api5ocki1;brow2cavalie0india2;bengaWre3;ds;arlotte horAicago 3;b4cubs,fire,wh3;iteB;ea0ulR;diff3olina panthe0; c3;ity;altimore 9lackburn rove0oston 5rooklyn 3uffalo bilN;ne3;ts;cel4red3; sox;tics;rs;oriol1rave2;rizona Ast8tlanta 3;brav1falco2h4u3;nited;aw9;ns;es;on villa,r3;os;c5di3;amondbac3;ks;ardi3;na3;ls",
@@ -48,13 +48,13 @@ export default {
   "LastName": "true¦0:9G;1:9W;2:9O;3:9Y;4:9I;5:8L;6:9L;7:A1;8:9F;9:8A;A:78;B:6G;C:6K;a9Vb8Nc7Ld6Ye6Tf6Fg5Wh59i55j4Qk45l3Nm2Sn2Fo27p1Oquispe,r18s0Ft05vVwOxNyGzD;aytsAEhD;aDou,u;ng,o;aGeun81iDoshiAAun;!lD;diDmaz;rim,z;maDng;da,guc98mo6VsDzaA;aAhiA8;iao,u;aHeGiEoDright,u;jc8Tng;lDmm0nkl0sniewsA;liA2s3;b0iss,lt0;a5Tgn0lDtanabe;k0sh;aHeGiEoDukB;lk5roby5;dBllalDnogr2Zr10ss0val37;ba,obos;lasEsel7P;lGn dFrg8FsEzD;qu7;ily9Pqu7silj9P;en b35ijk,yk;enzue96verde;aLeix1KhHi2j6ka3IoGrFsui,uD;om50rD;c2n0un1;an,embl8UynisA;dor96lst31m4rr9th;at5Ni7NoD;mErD;are70laci65;ps3s0Z;hirBkah8Enaka;a01chXeUhQiNmKoItFuEvDzabo;en8Bobod34;ar7bot4lliv2zuA;aEein0oD;i68j3Myan8W;l6rm0;kol5lovy5re6Rsa,to,uD;ng,sa;iDy60;rn5tD;!h;l5ZmEnDrbu;at8gh;mo6Eo6K;aFeDimizu;hu,vchD;en7Duk;la,r17;gu8mDoh,pulve8Trra4S;jDyD;on5;evi6Giltz,miDneid0roed0ulz,warz;dEtD;!z;!t;ar42h6ito,lFnDr4saAto,v4;ch7d0AtDz;a4Pe,os;as,ihBm3Zo0Q;aOeNiKoGuEyD;a67oo,u;bio,iz,sD;so,u;bEc7Bdrigue57g03j73mDosevelt,ssi,ta7Nux,w3Z;a4Ce0O;ertsDins3;!on;bei0LcEes,vDzzo;as,e8;ci,hards3;ag2es,it0ut0y9;dFmEnDsmu7Zv5F;tan1;ir7os;ic,u;aSeLhJiGoErDut6;asad,if60ochazk1V;lishc23pDrti63u55we67;e2Tov48;cEe09nD;as,to;as61hl0;aDillips;k,m,n5L;de3AetIna,rGtD;ersErovDtersC;!a,ic;en,on;eDic,ry,ss3;i8ra,tz,z;ers;h71k,rk0tEvD;ic,l3T;el,t2O;bJconnor,g2ClGnei5QrEzD;demir,turk;ella3MtDwe5O;ega,iz;iDof6GsC;vDyn1E;ei8;aPri1;aLeJguy1iFoDune44ym2;rodahl,vDwak;ak3Uik5otn57;eEkolDlsCx3;ic,ov6X;ls1miD;!n1;ils3mD;co42ec;gy,kaEray2varD;ro;jiDmu8shiD;ma;aWcUeQiPoIuD;lGnFrDssoli5T;atDpTr68;i,ov4;oz,te4C;d0l0;h2lInr13o0GrEsDza0Y;er,s;aFeEiDoz5r3Ete4C;!n6F;au,i8no,t4N;!l9;i2Rl0;crac5Ohhail5kke3Qll0;hmeFij0j2FlEn2Xrci0ssiDyer19;!er;n0Io;dBti;cartDlaughl6;hy;dMe6Egnu5Fi0jer35kLmJnci5ArFtEyD;er,r;ei,ic,su1O;iEkBqu9roqu6tinD;ez,s;a55c,nD;!o;a53mD;ad5;e5Pin1;rig4Ps1;aSeMiIoGuEyD;!nch;k4nDo;d,gu;mbarDpe2Svr4;di;!nDu,yana1T;coln,dD;bDholm;erg;bed5UfeGhtFitn0kaEn6rDw2H;oy;!j;in1on1;bvDvD;re;iDmmy,rsCu,voie;ne,t12;aTennedy,h2iSlQnez48oJrGuEvar2woD;k,n;cerDmar59znets5;a,o2H;aDem0i31yeziu;sni3RvD;ch3W;bay4Grh0Ksk0UvaFwalDzl5;czDsA;yk;cFlD;!cDen3S;huk;!ev4ic,s;e6uiveD;rt;eff0l4mu8nnun1;hn,llFminsArEstra33to,ur,yDzl5;a,s0;j0HlsC;oe;aMenLha2Qim0RoEuD;ng,r4;e2KhFnErge2Ku2OvD;anB;es,ss3;anEnsD;en,on,t3;nesDsC;en,s1;ki27s1;cGkob3RnsDrv06;en,sD;enDon;!s;ks3obs1;brahimBglesi3Ake4Ll0DnoZoneFshikEto,vanoD;u,v4A;awa;scu;aPeIitchcock,jaltal6oFrist46uD;!aDb0gh9ynh;m2ng;a24dz4fEjga2Tk,rDx3B;ak0Yvat;er,fm3B;iGmingw3NnErD;nand7re8;dDriks1;ers3;kkiEnD;on1;la,n1;dz4g1lvoLmJnsCqIrr0SsFuEyD;as36es;g1ng;anEhiD;mo0Q;i,ov08;ue;alaD;in1;rs1;aNeorgMheorghe,iKjonJoGrEuDw3;o,staf2Utierr7zm2;ayDg4iffitVub0;li1H;lub3Rme0JnEodD;e,m2;calv9zale0H;aj,i;l,mDordaL;en7;iev3A;gnJlGmaFnd2Mo,rDs2Muthi0;cDza;ia;ge;eaElD;agh0i,o;no;e,on;ab0erLiHjeldsted,lor9oFriedm2uD;cDent9ji3E;hs;ntaDrt6urni0;na;lipEsD;ch0;ovD;!ic;hatBnanFrD;arDei8;a,i;deS;ov4;dGinste6riksCsDva0D;cob2YpDtra2W;inoza,osiL;en,s3;er,is3wards;aUeMiKjurhuJoHrisco0ZuEvorakD;!oQ;arte,boEmitru,rDt2U;and,ic;is;g2he0Imingu7n2Ord1AtD;to;us;aDmitr29ssanayake;s,z; GbnaFlEmirDrvis1Lvi,w2;!ov4;gado,ic;th;bo0groot,jo04lEsilDvri9;va;a cruz,e3uD;ca;hl,mcevsAnEt2EviD;d5es,s;ieDku1S;ls1;ki;a06e01hOiobNlarkMoFrD;ivDuz;elli;h1lHntGoFrDs26x;byn,reD;a,ia;ke,p0;i,rer0N;em2liD;ns;!e;anu;aLeIiu,oGriDuJwe;stD;eDiaD;ns1;i,ng,uFwDy;!dhury;!n,onEuD;ng;!g;kEnDpm2tterjee,v7;!d,g;ma,raboD;rty;bGl08ng4rD;eghetEnD;a,y;ti;an,ota0L;cer9lder3mpbeIrFstDvadi07;iDro;llo;doEt0uDvalho;so;so,zo;ll;es;a08eWhTiRlNoGrFyD;rne,tyD;qi;ank5iem,ooks,yant;gdan5nFruya,su,uchEyHziD;c,n5;ard;darDik;enD;ko;ov;aEondD;al;nco,zD;ev4;ancRshwD;as;a01oDuiy2;umDwmD;ik;ckNethov1gu,ktLnJrD;gGisFnD;ascoDds1;ni;ha;er,mD;ann;gtDit7nett;ss3;asD;hi;er,ham;b4ch,ez,hMiley,kk0nHrDu0;bEnDua;es,i0;ieDosa;ri;dDik;a8yopadhyD;ay;ra;er;k,ng;ic;cosZdYguilXkhtXlSnJrGsl2yD;aEd6;in;la;aEsl2;an;ujo,ya;dFgelD;ovD;!a;ersGov,reD;aDjL;ss1;en;en,on,s3;on;eksejGiyGmeiFvD;ar7es;ez;da;ev;ar;ams;ta",
   "MaleName": "true¦0:DN;1:CO;2:D6;3:AJ;4:CK;5:BZ;6:CF;7:D2;8:BS;9:AR;A:DA;B:D3;C:94;D:BM;aC9bB7cA7d98e8If82g7Fh6Si6Cj5Ek52l4Fm37n2Uo2Op2Gqu2Er1Ms12t0Gu0Fv08wUxTyJzE;aEor0;cEh9Jkaria,n0C;hFkE;!aC7;ar5UeC6;aMoGuE;sEu2LvBJ;if,uf;nGsFusE;ouf,sE;ef;aEg;s,tE;an,h0;hli,nB8ssY;avi3ho4;aNeLiGoEyaBN;jcie87lfgang,odrow,utE;!er;lEnst1;bGey,fredAlE;aAZiE;am,e,s;e97ur;i,nde9sE;!l8t1;lFyE;l1ne;lEt3;a9Xy;aHiEladimir,ojte7U;cFha0kt67nceErgA5va0;!nt;e3Xt65;lentEn9S;inE;!e;ghBElyss59nax,sm0;aXeShOiMoIrGuFyE;!l3ro6s1;n7r59;avAHeEist0oy,um0;ntA9v5Wy;bGd8RmEny;!as,mEoharu;aCBie,y;iCy;mEt5;!my,othy;adGeoFia0KomE;!as;!do8G;!de5;dHrE;en98rE;an97eEy;ll,n96;!dy;dgh,ha,iEnn3req,tsu4R;cAPka;aUcotSeQhMiKoIpenc3tEur1Xylve96zym1;anGeEua85;f0phBCvEwa84;e5Zie;!islaw,l8;lom1uE;leyma6ta;dElCm1yabonga;!dhart74n8;aGeE;lErm0;d1t1;h7Kne,qu11un,wn,y6;aEbasti0k2Cl4Prg4Mth,ymoAE;m5n;!tE;!ie,y;lFmEnti2Gq58ul;!ke5JmDu4;ik,vato7O;aZeVhe9ViRoIuFyE;an,ou;b7DdFf5pe7KssE;!elBI;ol3Fy;an,bLc62dJel,geIh0landAmHnGry,sFyE;!ce;coe,s;!aA1nD;an,eo;l45r;er78g3n8olfo,riE;go;bDeAQ;cEl8;ar6Ic6HhFkEo;!ey,ie,y;a8Vie;gFid,ubByEza;an1KnZ;g9SiE;na9Os;ch6Qfa4lImHndGpha4sFul,wi2HyE;an,mo6U;h7Jm5;alAWol2Uy;iACon;f,ph;ent2inE;cy,t1;aJeHhilGier6TrE;aka18eE;m,st1;!ip,lip;dA4rcy,tE;ar,e3Er1Y;b4Hdra73tr6JulE;!o19;ctav3Di3liv3m9Yndrej,rIsFtEum7wB;is,to;aFc7k7m0vE;al5S;ma;i,vM;aMeKiGoEu38;aEel,j5l0ma0r3I;h,m;cFg4i46kEl2R;!au,h7Gola;hEkEolC;olC;al,d,il,ls1vE;il8J;hom,tE;e,hE;anEy;!a4i4;a00eXiNoIuFyE;l2Gr1;hamFr6KstaE;fa,p54;ed,mI;di0Xe,hamGis2CntFsEussa;es,he;e,y;ad,ed,mE;ad,ed;cJgu4hai,kHlGnFtchE;!e9;a7Uik;house,o0Ct1;ae5Oe9MolE;aj;ah,hE;aFeE;al,l;el,l;hFlv2rE;le,ri9v2;di,met;ay0hUjd,ks2AlSmadXnRrLs1tGuricFxE;imilianAwe9;e,io;eHhFiCtEus,yC;!eo,hew,ia;eEis;us,w;j,o;cIio,kHlGqu6Ysha9tEv2;iEy;!m,n;in,on;el,oQus;!el90oPus;iHu4;achEcolm,ik;ai,y;amFdi,eEmoud;sh;adEm5G;ou;aXeRiPlo39oLuFyE;le,nd1;cHiGkEth3uk;aEe;!s;gi,s,z;as,iaE;no;g0nn7BrenGuEv81we9;!iE;e,s;!zo;am,oE;n4r;a7Uevi,la4AnIonHst3thaGvE;eEi;nte;bo;!a6Del;!ny;mFnd1rEur54wr54;ry,s;ar,o4Y;aMeIhal7GiFristEu4Ky6J;i0o54;er0p,rE;k,ollE;os;en0iGnErmit,v3U;!dr3XnEt1;e18y;r,th;cp3j5m5Sna6OrFsp7them,uE;ri;im,l;a01eViToHuE;an,lEst2;en,iE;an,en,o,us;aOeMhnLkubCnJrHsE;eFhEi7Vue;!ua;!ph;dEge;i,on;!aEny;h,s,th55;!ath54ie,nD;!l,sEy;ph;o,qu2;an,mE;!mD;d,ffHrEs5;a5YemFmai6oEry;me,ni0Y;i7Fy;!e5OrE;ey,y;cLdBkJmIrGsFvi3yE;dBs1;on,p3;ed,od,rEv4V;e5Bod;al,es4Mis1;a,e,oEub;b,v;ob,quE;es;aXbRchiQgOkeNlija,nuMonut,rKsGtEv0;ai,suE;ki;aFha0i6ZmaEsac;el,il;ac,iaE;h,s;a,vinEw2;!g;k,nngu5F;!r;nacEor;io;ka;ai,rahE;im;aQeKoJuEyd7;be2FgHmber4KsE;eyFsE;a2e2;in,n;h,o;m3ra36sse2wa40;aIctHitHnrFrE;be28m0;iEy;!q0Z;or;th;bMlLmza,nKo,rGsFyE;a47dB;an,s0;lGo4Nry,uEv8;hi44ki,tE;a,o;an,ey;k,s;!im;ib;aWeSiQlenPoMrIuE;ilFsE;!tavo;herme,lerE;mo;aGegEov3;!g,orE;io,y;dy,h5J;nzaFrE;an,d1;lo;!n;lbe4Xno,oE;rg37van4X;oGrE;aEry;ld,rdA;ffr8rge;brFlBrEv2;la14r3Hth,y;e33ielE;!i5;aSePiNlLorrest,rE;anFedEitz;!dDer11r11;cGkE;!ie,lE;in,yn;esLisE;!co,z2W;etch3oE;yd;d4lEonn;ip;deriFliEng,rnan05;pe,x;co;bi0di,hd;dYfrXit0lSmLnIo2rGsteb0th0uge6vEymBzra;an,eE;ns,re2X;gi,i0AnErol,v2w2;estAie;oFriqEzo;ue;ch;aJerIiFmE;aIe2Q;lErh0;!iE;o,s;s1y;nu4;be0Bd1iGliFm3t1viEwood;n,s;ot1Ss;!as,j4EsE;ha;a2en;!d2Vg7mHoFuFwE;a26in;arE;do;oWuW;a02eRiPoHrag0uGwFylE;an,l0;ay6ight;a6dl8nc0st2;minHnFri0ugEvydCy29;!lC;!a2HnEov0;e9ie,y;go,iFykC;as;cEk;!k;armuEll1on,rk;id;andNj0lbeMmetri5nKon,rIsGvFwExt3;ay6ey;en,in;hawn,moE;nd;ek,rE;ick;is,nE;is,y;rt;re;an,le,mLnKrGvE;e,iE;!d;en,iGne9rEyl;eEin,yl;l35n;n,o,us;!i4ny;iEon;an,en,on;a08e06hYiar0lOoJrHuFyrE;il,us;rtE;!is;aEistob0S;ig;dy,lHnFrE;ey,neli5y;or,rE;ad;by,e,in,l2t1;aIeGiEyK;fEnt;fo0Et1;meEt5;nt;rGuFyE;!t1;de;enE;ce;aIeGrisE;!toE;ph3;st3;er;d,rEs;b4leE;s,y;cEdric,s7;il;lHmer1rE;ey,lFro9y;ll;!os,t1;eb,v2;a07eZiVlaUoSrFuEyr1;ddy,rtL;aMeHiGuFyE;an,ce,on;ce,no;an,ce;nFtE;!t;dFtE;!on;an,on;dFndE;en,on;!foEl8y;rd;bby,rEyd;is;i6ke;bGlFshE;al;al,lD;ek;nIrEshoi;at,nFtE;!r1B;aEie;rdA;!iFjam2nD;ie,y;to;kaNlazs,nIrE;n8rEt;eEy;tt;ey;dEeF;ar,iE;le;ar16b0Ud0Qf0Ogust2hm0Li0Ija0Hl03mZnSputsiRrIsaHugust5veFyEziz;a0kh0;ry;us;hi;aLchKiJjun,maInGon,tEy0;hEu09;ur;av,oE;ld;an,ndA;!el,ki;ie;ta;aq;as,dIgelAtE;hony,oE;i6nE;!iAy;ne;er,reEy;!as,i,s,w;iGmaEos;nu4r;el;ne,r,t;an,beQdBeKfIi,lHonGphYt1vE;aOin;on;so,zo;an,en;onUrE;ed;c,jaHksandGssaHxE;!andE;er,ru;ar,er;ndE;ro;rtA;ni;dBm7;ar;en;ad,eE;d,t;in;onE;so;aFi,olfAri0vik;!o;mEn;!a;dIeHraFuE;!bakr,lfazl;hEm;am;!l;allJelGoulaye,ulE;!lErG;ah,o;! rE;ahm0;an;ah;av,on",
   "Person": "true¦ashton kutchXbWcRdPeNgastShLinez,jKkJleImEnDoCpAr5s4t2va1w0;arrGednesday adams,oode;lentino rossi,n go4;a0heresa may,iger woods,yra banks;tum,ylor;addam hussain,carlett johanssNlobodan milosevic;ay romano,e3o1ush limbau0;gh;d stewart,nald0;inho,o;ese witherspoIil6;a0ipM;lmLris hiltG;prah winfrIra;ettAiles crane;essia2i0ubarek;randa ju0tt romnF;ly;en;bron james,e;anye west,endall,iefer sutherland,obe bryant;aime,effers7k rowling;a0itlBulk hogan;lle berry,rris5;ff0meril lagasse,zekiel;ie;a0enzel washingt2ick wolf;lt1nte;ar1lint0;on;dinal wols1son0;! palm2;ey;arack obama,rock;er",
-  "Adjective": "true¦0:9F;1:89;2:85;3:96;4:90;5:83;6:8L;7:8E;8:91;9:8O;A:7R;B:5Y;C:95;D:7C;a7Sb7Ac6Td6Ce5Uf59g51h4Qi3Rjuni46k3Pl3Fm33n2To2Ep1Zquart67r1Ps0Tt0LuPvLwEye1L;ast55eIholeHiGoE;man5oErthwhi6u0H;dAzy;despr8Qn w74s6J;!sa6;ather14eEll o5Mste2Q;!k5;aGeFiEola5G;b9Ice versa,gi2Qle;ng4Wrsa5E;ca0lu8M;lt08nKpGrFsEttermo9A;efDu4;b6Cge0; Gb2ApFsEti89;ca6et,ide d3N;er,i4M;f3Vto da3;aYbecom2cXdSeRfPiOkn3KmNpMrJsFtoIus1wE;a08iel4F;eAi2FoGpFuE;pervis1spect2;e0okA;ld;eEu5;cognSgul0MlEsolv1;at1ent2;a9recede00;arri1et;que,vers4;air,orE;eseAtun7J;mploy1nd2xpect1;eEue;cid1rE;!a6XcovBly2sIwE;aEei7N;tBy;heck1onvinc2;ppeal2ssum2tteFuthorE;iz1;nd1;im48ra;aJeGhough5Fip 1QoFrE;anspa7Ai3;gethBle8Hp notch,rp9;ena6MmpFrE;r3Gtia6Z;e8o6X;leEst3S;nt1;a08c04e01h00iYkiXmug,nobbi43oSpPqueami43tJuEymb7C;bGi generis,pEr5;erEre60;! dupBb,viZ;du1p49sEurb56;eq7Itanda83;atu6GeIi0VrEy3W;aEin4I;ightEy; fEfE;or64;adfa7Vri6;arFeEirit1lend9ot on;c2Ze36;k5se; caIlub6mbBphisticHrGuEviS;ndEth0Z;proof;d9ry;at1;ll1;g1ZnE;ce5Eg6;am33eB;at1co1Jem5lfFnEre7;so66; suf42i3D;aFholEient2L;ar5;rlEth2;et;cr1me,tisfac5R;aMeHheumato9iFoE;bu79ttAy4;ghtEv4;-w2fD;cHdu70lGnown1sEtard1;is3GoE;lu3na0;e1Auc3E;e0ondi3;b9ci4;aPeNicayu7laMopuli70rFuE;bl66njabi;eJiHoE;b68fFmi3FpEvAx20;er,ort6E;a7u6H;maEor,sti7va3;!ry;ci6Eexist2ma1Rpa9;c9id;ac2ArEti3;feCma37ti37v6A;i2ArtFsE;s65t;-4Hi4;bPffNkMld LnJrIthButHverE;!aFni5Mseas,t,wE;ei5Lrou5L;ll;do11er,si4T;d31g1Q; bEbEgo2li7;oa6G;fashion1school;!ay; gua6DbEli7;eat;eFsE;ce7er0Fo0W;dia0se;aMeLiKoEuanc1; mo3ZnGrthEt5G;!eE;rn;chaFdescri61prof2GsE;top;la0;ght5;arby,cessa4Nighbor5xt;k1usiat2;aLeKiIoFuE;d18ltip6;deFl17nEot,st;ochro3Pth5;rn,st;ddle ag1nE;dblUi;nac2re;cGgenta,in,j08keshift,mmFnEscu4S;da43y;ali2Qoth;ab3Fho;aMeJiHoEumber2;ngEuti1G;stand2tE;erm,i3D;ghtwei4IteraE;l,te;ft-w2gEssBth4;al,eEit0G;nda41;ngu9ps1te5;aput,ind5nE;ow2;gno5Dll08m00nGpso 2DrE;a3releE;va0; VaUcoRdOe2GfNgrMhibi3Li01nLoKsGtEvalu4G;aCeE;n4Prdep22;a7iFolEuboH;ub6ve0;de,gE;nifica0;rdi4H;a3er;own;eriFluenRreq4C;eFiHoE;or;fini3p1Rtermi4B;mpFnside8rE;reC;le3;ccu8deq4Cppropr3J;fEsitu,vitro;ro0;mIpE;arGeFl0XoEropB;li3r0U;nd2rfeC;ti4;aFeEi0Y;d3An41;tu28;egFiE;c0Qte8;al,itE;ima3;aLelKiIoEuma7;meGnFrrEs0Aur5;if3E;e46o2U; ma2Hsick;ghfalut1QspE;an3A;liZpfD;i2QlFnErd04tL;dy;f,low1;aiKener35iga2Dlob4oJraGuE;ilEng ho;ty;cFtE;efDis;efD;ldAod;nfD;aUeRinQlOoHrE;aFeEozA;q35tfD;gi6nt2U;oIrE; keeps,eGge0MmBtunEwa3N;ateE;!ly;go2i1Is2S;liE;sh;ag3KuE;e0oresce0;al,i3;dFmini7rE;ti6; up;bl1i0l2VmiliGr Eux;oEreach2;ff;ar;aSfRlPmMnKqu4reCthere4veryday,xE;aCem33plHquisi3traGuE;be38lE;ta0;!va1O;icE;it; Eti0S;rou3sui3;erFiE;ne0;ge0;dEe1G;er5;ficie0;gBsE;t,ygo2;er;aTeKiFoErea1Cue;mina0ne,rma0ubN;dact1Sfficult,m,sFverE;ge0se;creFeMjoi0pa8tE;a0inC;et,te; LadpKceJfiIgene8liGpFreliCspe8voE;id,ut;ende0;ca3ghE;tfD;a0ni3;as1;an;facto;i5ngeroZ;arTeRivil,oGrEustoma0V;aEu0Rystal19;vAz1;erNgniza0loMmLnGrFveE;rt;po8ru20;cHduKgr1HjGsFtraE;dic0Fry;eq1Fta0;oi0ug4;a19i1I;mensu8pass1D;ni4ss4;ci16;leba3rtaE;in;diac,efD;aTeMiIliRoFrEuck nak1;and new,isk,on1S;gFldface,naE; fiYfiY;us;gFzarE;re;tiE;me;autifDhiIloHnFsEyoI;iRt;eZiEt;gn;v1w;nd;ul;ckwards,nkru18rE;e,rA; priori,b10c0Xd0Rf0Lg0Gh0Fl08mp6nt05pXrQsOttracti0KuKvHwE;aFkE;wa17;ke,re;ant garFeraE;ge;de;diGtE;heEoimmu7;nt05;toM;hAleep;en;bitIchiv4roHtE;iEsy;fiE;ci4;ga0;raE;ry;pEt;aJetiz2roE;prGximE;ate5;ly;ia3;ing;re0;iquEsy;at1e;ed;cohIiPkaGl,oFriEterV;ght;ne,of;li7;ne;olE;ic;ead;ain04ed,gressiHrE;eeE;ab6;le;ve;fFra9;id;ectFlE;ue0;ioE;na3; IeGvE;erE;se;pt,qE;ua3;hoc,infinitum;cu8tu4u3;al;ra3;erPlNoLrIsFuE;nda0;e0olu3traC;ct;te;eaFuE;pt;st;aEve;rd;aEe;ze;ra0;nt",
-  "Adj|Past": "true¦0:2Z;1:2Q;2:2T;a2Ib2Cc1Td1Ae14f0Zgift0h0Wi0Pj0Oknown,l0Lm0Gn0Eo0Bp04qua03rUsEtAu7v5w3;arp0ea3or5;kIth2S;a3e0V;ri0;n3pd1s0;derstood,i3;fi0t0;ar5hreatCr3wi2Q;a3ou18;ck0in0pp0;get0ni1M;aGcaFeEhDimCm00oak0pAt6u3;bsid27gge2Ls3;pe3ta1R;ct0nd0;at0e5r3uU;ength3ip0;en0;am0reotyp0;eci3ik0ott0;al1Zfi0;pHul1;ar0ut;al0c1Ile2t1R;r0tt25;t3ut0;is1Jur1;aAe3;c7duc0f1Cg6l1new0qu5s3;pe2t3;or0ri2;e22ir0;ist1Xul1;eiv0o3;mme09rd0v1V;lli0ti1A;li19;arallel0i0Kl7o6r3ump0;e4o3;ce0Ilo0Hnou1Tpos0te2;fe0Loc8pY;i1Gli0R;a3e19;nn0;c4rgan1Bverlo3;ok0;cupi0;e3ot0;ed0gle2;a5e4ix0o3ut0;di0Vt0G;as0Qlt0;n3rk0;ag0ufact0O;eft,i4o3;ad0st;cens0mit0st0;agg0us0N;mp8n3sol1;br0debt0f6t3volv0;e3ox0F;gr1n3re18;d0si0J;e2l1oX;li0oLrov0;amm10e1o3;ok0r3;ri0E;aNe6i5lavo09oc05r3;a3i0;ct07g0Mm0;niXx0;ar0;duc1l1mbarraKn7quipp0stabliVx3;agger1p3te5;a4e3;ct0rie0S;nd0;ha0QsZ;aJeAi3;gni01miniOre2s3;a7c5grun05t3;o3reBurb0;rt0;iplSou3;nt0rE;bl0;cenVdOf8l7pre6ra5t3velop0;a3ermO;il0;ng0;ss0;ay0ight0;e4o3;rm0;rr0;m3t0;ag0;alcul1eHharg0lGo8r5u3;lt3stomS;iv1;a4owd0u3;sh0;ck0mp0;d0lo9m6n3ok0vX;centr1f4s3troll0;idVolid1;us0;b4pl3;ic1;in0;ur0;assi5os0;lebr1n5r3;ti3;fi0;tralB;a7i6o4roken,urn3;ed,t;il0r0t3und;tl0;as0;k0laIs0;bandon0cJdGffe2lDnBpp9ss7u3ward0;g4thor3;iz0;me3;nt0;o5u3;m0r0;li0re3;ci1;im1ticip1;at0;leg0t3;er0;ct0;ju4o6va3;nc0;st0;ce3knowledg0;pt0;ed",
+  "Adjective": "true¦0:9C;1:86;2:82;3:93;4:8X;5:80;6:8I;7:8B;8:8Y;9:8L;A:7O;B:5W;C:92;D:79;E:7V;a7Qb78c6Sd6Be5Tf58g50h4Qi3Rjuni46k3Pl3Fm33n2To2Ep1Zquart66r1Ps0Ut0MuQvMwFye1L;ast54eJholeIiHoF;man5oFrthwhi6u0I;dAzy;despr8On w72s6I;!sa6;ather14eFll o5Lste2Q;!k5;aHeGiFola5F;b9Gce versa,gi2Qle;ng4Vrsa5D;ca0lu8K;lt09nLpHrGsFttermo98;efDu4;b6Bge0; Hb2ApGsFti87;ca6et,ide d3N;er,i4L;f3Vto da3;aZbecom2cYdTeSfQiPkn3KmOpNrKsGtoJus1wF;a09iel4F;eAi2FoHpGuF;pervis1spect2;e0okA;ld;eFu5;cognTgul0MlFsolv1;at1ent2;a9recede01;arri1et;que,vers4;air,orF;eseAtun7H;mploy1nd2xpect1;eFue;cid1rF;!a6VcovBly2sJwF;aFei7L;tBy;heck1onvinc2;ppeal2ssum2tteGuthorF;iz1;nd1;im47ra;aKeHhough5Eip 1QoGrF;anspa78i3;gethBle8Fp notch,rp9;ena6KmpGrF;r3GtiaE;e8o6V;leFst3R;nt1;a08c04e01h00iYkiXmug,nobbi42oSpPqueami42tKuFymb7A;bHi generis,pFr5;erFre5Y;! dupBb,viZ;du1p48sFurb55;eq7Gtanda81;atu6EeIi0VrFy3V;aightFin4H; fFfF;or63;adfa7Uri6;arGeFirit1lend9ot on;c30e37;k5se; caJlub6mbBphisticIrHuFviT;ndFth10;proof;d9ry;at1;ll1;g20nF;ce5Dg6;am33eB;co1Kem5lfGnFre7;soE; suf42i3D;aGholFient2M;ar5;rlFth2;et;cr1me,tisfac5Q;aNeIheumato9iGoF;bu78ttAy4;ghtFv4;-w2fD;cIdu6ZlHnown1sFtard1;is3GoF;lu3na0;e1Buc3E;e0ondi3;b9ci4;aQeOicayu7laNopuli6ZrGuF;bl65njabi;eKiIoF;b67fGmi3FpFvAx21;er,ort6D;a7u6G;maFor,sti7va3;!ry;ci6Dexist2ma1Spa9;c9id;ac2ArFti3;feCma37ti37v69;i2ArtGsF;s64t;-4Gi4;bQffOkNld MnKrJthButIverF;!aGni5Lseas,t,wF;ei5Krou5K;ll;do12er,si4S;d31g1R; bFbFgo2li7;oa6F;fashion1school;!ay; gua6CbFli7;eat;eGsF;ce7er0Go0X;dia0se;aNeMiLoFuanc1; mo3YnHrthFt5F;!eF;rn;chaGdescri60prof2GsF;top;la0;ght5;arby,cessaEighbor5xt;k1usiat2;aMeLiJoGuF;d19ltip6;deGl18nFot,st;ochro3Oth5;rn,st;ddle ag1nF;dblVi;nac2re;cHgenta,in,j09keshift,mmGnFscu4R;da42y;ali2Qoth;ab3Eho;aNeKiIoFumber2;ngFuti1G;stand2tF;erm,i3C;ghtwei4HteraF;l,te;ft-w2gFssBth4;al,eFit0H;ndaE;ngu9ps1te5;aput,ind5nF;ow2;gno5Cll09m01nHpso 2DrF;a3releF;va0; WaVcoSdPe2GfOgrNhibi3Ki02nMoLsHtFvalu4F;aCeF;n4Ordep22;a7iGolFuboI;ub6ve0;de,gF;nifica0;rdi4G;a3er;own;eriGluenSreq4B;eGiIoF;or;fini3p1Rtermi4A;mpGnside8rF;reC;le3;ccu8deq4Bppropr3I;fFsitu,vitro;ro0;mJpF;arHeGl0XoFropB;li3r0U;nd2rfeC;ti4;aGeFi0Y;d39n40;tu27;egGiF;c0Qte8;al,itF;ima3;aMelLiJoFuma7;meHnGrrFs0Aur5;if3D;e45o2T; ma2Gsick;ghfalut1PspF;an39;liZpfD;iEllow1nFrd04tL;dy;aiLener35iga2Dlob4oKraHuF;ilFng ho;ty;cGtF;efDis;efD;ldAod;nfD;aVeSinRlPoIrF;aGeFozA;q35tfD;gi6nt2U;oJrF; keeps,eHge0NmBtunFwa3N;ateF;!ly;go2i1Is2S;liF;sh;ag3KuF;e0oresce0;al,i3;dGmini7rF;ti6; up;bl1l2VmiliHr Fux;oFreach2;ff;ar;aTfSlQmNnLqu4reCthere4veryday,xF;aCem33plIquisi3traHuF;be38lF;ta0;!va1O;icF;it; Fti0S;rou3sui3;erGiF;ne0;ge0;dFe1G;er5;ficie0;gBsF;t,ygo2;er;aUeLiGoFreaEue;mina0ne,rma0ubO;dact1Sfficult,m,sGverF;ge0se;creGeNjoi0pa8tF;a0inC;et,te; MadpLceKfiJgene8liHpGreliCspe8voF;id,ut;ende0;ca3ghF;tfD;a0ni3;as1;an;facto;i5ngeroZ;arTeRivil,oHrFustomaE;aFu0Rystal19;vAz1;erNgniza0loMmLnGrpo8veF;rt;cIduLgr1IjHsGtraF;dic0Gry;eq1Gta0;oi0ug4;a1Ai1J;mensu8pass1E;ni4ss4;ci17;leba3rtaF;in;diac,efD;aUeNiJliSoGrFuck nak1;and new,isk,on1T;gGldface,naF; fiZfiZ;us;gGzarF;re;tiF;me;autifDhiJloInGsFyoJ;iSt;e00iFt;gn;v1w;nd;ul;ckwards,rF;e,rA; priori,b11c0Yd0Sf0Mg0Hh0Gl09mp6nt06pYrRsPttracti0LuLvIwF;aGkF;wa18;ke,re;ant garGeraF;ge;de;diHtF;heFoimmu7;nt06;toE;hAleep;en;bitJchiv4roItF;iFsy;fiF;ci4;ga0;raE;ry;pFt;aKetiz2roF;prHximF;ate5;ly;ia3;ing;re0;iquFsy;at1e;ed;cohJiQkaHl,oGriFterW;ght;ne,of;li7;ne;olF;ic;ead;ain05ed,gressiIrF;eeF;ab6;le;ve;fGra9;id;ectGlF;ue0;ioF;na3; JeHvF;erF;se;pt,qF;ua3;hoc,infinitum;cu8tu4u3;al;ra3;erQlOoMrJsGuF;nda0;e0olu3traC;ct;te;eaGuF;pt;st;aFve;rd;aFe;ze;ra0;nt",
+  "Adj|Past": "true¦0:30;1:2R;2:2U;a2Jb2Dc1Ud1Be15f10gift0h0Xi0Qj0Pknown,l0Mm0Hn0Fo0Cp05qua04rVsEtAu7v5w3;arp0ea3or5;kIth2T;a3e0W;ri0;n3pd1s0;derstood,i3;fi0t0;ar5hreatCr3wi2R;a3ou19;ck0in0pp0;get0ni1N;aHcaGeEhDimCm01oak0pAt6u3;bsid28gge2Ms3;pe3ta1S;ct0nd0;at0e5r3uV;ength3ip0;en0;am0reotyp0;eci3ik0ott0;al20fi0;pIul1;ar0ut;a3c1Jle2t1S;l0t0;r0tt25;t3ut0;is1Jur1;aAe3;c7duc0f1Cg6l1new0qu5s3;pe2t3;or0ri2;e22ir0;ist1Xul1;eiv0o3;mme09rd0v1V;lli0ti1A;li19;arallel0i0Kl7o6r3ump0;e4o3;ce0Ilo0Hnou1Tpos0te2;fe0Loc8pY;i1Gli0R;a3e19;nn0;c4rgan1Bverlo3;ok0;cupi0;e3ot0;ed0gle2;a5e4ix0o3ut0;di0Vt0G;as0Qlt0;n3rk0;ag0ufact0O;eft,i4o3;ad0st;cens0mit0st0;agg0us0N;mp8n3sol1;br0debt0f6t3volv0;e3ox0F;gr1n3re18;d0si0J;e2l1oX;li0oLrov0;amm10e1o3;ok0r3;ri0E;aNe6i5lavo09oc05r3;a3i0;ct07g0Mm0;niXx0;ar0;duc1l1mbarraKn7quipp0stabliVx3;agger1p3te5;a4e3;ct0rie0S;nd0;ha0QsZ;aJeAi3;gni01miniOre2s3;a7c5grun05t3;o3reBurb0;rt0;iplSou3;nt0rE;bl0;cenVdOf8l7pre6ra5t3velop0;a3ermO;il0;ng0;ss0;ay0ight0;e4o3;rm0;rr0;m3t0;ag0;alcul1eHharg0lGo8r5u3;lt3stomS;iv1;a4owd0u3;sh0;ck0mp0;d0lo9m6n3ok0vX;centr1f4s3troll0;idVolid1;us0;b4pl3;ic1;in0;ur0;assi5os0;lebr1n5r3;ti3;fi0;tralB;a7i6o4roken,urn3;ed,t;il0r0t3und;tl0;as0;k0laIs0;bandon0cJdGffe2lDnBpp9ss7u3ward0;g4thor3;iz0;me3;nt0;o5u3;m0r0;li0re3;ci1;im1ticip1;at0;leg0t3;er0;ct0;ju4o6va3;nc0;st0;ce3knowledg0;pt0;ed",
   "Determiner": "true¦aBboth,d9e6few,l4mu8neiDplenty,s3th2various,wh0;at0ich0;evC;at,e4is,ose;everal,ome;a,e0;!ast,s;a1i6l0very;!se;ch;e0u;!s;!n0;!o0y;th0;er",
-  "Adj|Present": "true¦a01blu00cSdNeLfIhollHidOlEmCnarrHoBp9qua8r7s4t2utt3w0;aJet,ound,ro0;ng,ug02;end0hin;er;e1l0mooth,our,pa8u8;i2ow;cu6daWlOpaK;eplicaVigW;ck;aEr0;esePime,ompt;bscu1pen,wn;atu0eModeE;re;a1eKi0;gOve;st,y;ow;i1r0;ee,inge;rm;l0mpty,xpress;abo4ic7;amp,e2i1oub0ry;le;ffu8r5;fu7libe0;raB;l4o0;mple9n2ol,rr1unterfe0;it;ect;juga6sum5;e1o0;se;an;nt;lig2pproxi0;ma0;te;ht",
+  "Adj|Present": "true¦a06b03cUdPeNfJhollIidQlEmCnarrIoBp9qua8r7s4t2utt3w0;aKet,ound,ro0;ng,ug07;end0hin;er;e1hort,l0mooth,our,pa8tray,u8;i2ow;cu6da01leRpaM;eplica00ig01;ck;aGr0;eseTime,omU;bscu1pen,wn;atu0e3odeG;re;a2e1i0;gTve;an;st,y;ow;a2i1oul,r0;ee,inge;rm;iHke,ncy,st;l0mpty,xpress;abo4ic7;amp,e2i1oub0ry,ull;le;ffu9re6;fu8libe0;raE;alm,l5o0;mpleCn3ol,rr1unterfe0;it;e0u7;ct;juga8sum7;ea1o0;se;n,r;ankru1lu0;nt;pt;lig2pproxi0;ma0;te;ht",
   "Adverb": "true¦a08b05d01eXfRhePinOjustNkinda,likewi00mLnIoDpBquite,r8s4t1up0very,well; to,wards5;h1iny bit,o0wiO;o,t6;en,us;eldom,o0uch;!me1rt0; of;hZtimes,w0A;a1e0;alT;ndomSthN;ar excellenEer0oint blank; Nhaps;f3n0;ce0ly;! 0;ag03moX; courIten;ewKo0; longEt 0;onIwithstanding;aybe,eanwhiAore0;!ovB;! aboV;deed,steW;lla,n0;ce;or2u0;lArther0;!moK; 0ev3;examp0good,suI;le;n1v0;er; mas0ough;se;e0irect1; 1finite0;ly;ju9trop;ackward,y 0;far,n0;ow; DbroCd nauseam,gBl6ny3part,s2t 0w4;be6l0mo6wor6;arge,ea5; soon,ide;mo1w0;ay;re;l 1mo0ready,so,ways;st;b1t0;hat;ut;ain;ad;lot,posteriori",
   "Currency": "true¦$,aud,bQcOdJeurIfHgbp,hkd,iGjpy,kElDp8r7s3usd,x2y1z0¢,£,¥,ден,лв,руб,฿,₡,₨,€,₭,﷼;lotyQł;en,uanP;af,of;h0t5;e0il5;k0q0;elK;oubleJp,upeeJ;e2ound st0;er0;lingG;n0soF;ceEnies;empi7i7;n,r0wanzaCyatC;!onaBw;ls,nr;ori7ranc9;!os;en3i2kk,o0;b0ll2;ra5;me4n0rham4;ar3;e0ny;nt1;aht,itcoin0;!s",
-  "Comparable": "true¦0:3B;1:3Q;2:3F;3:2D;a3Ub3Cc30d2Qe2Jf27g1Vh1Li1Fj1Ek1Bl14m0Yn0To0Sp0Jqu0Hr08sJtEuDvBw5y4za0R;el11ou3A;a8e6hi1Hi4ry;ck0Dde,l4n1ry,se;d,y;a4i3T;k,ry;nti34ry;a4erda2ulgar;gue,in,st;g0pcomi31;a7en2Thi6i5ough,r4;anqu28en1ue;dy,g36me0ny,r03;ck,rs24;ll,me,rt,wd3I;aRcarQePhNiMkin0BlImGoEpDt7u5w4;eet,ift;b4dd0Vperfi1Wrre24;sta22t3;a8e7iff,r5u4;pUr1;a4ict,o2P;ig2Wn0N;a1ep,rn;le,rk,te0;e1Oi2Wright0;ci1Vft,l4on,re;emn,id;a4el0;ll,rt;e6i4y;g2Nm4;!y;ek,nd2T;ck,l0mp3;a4iRort,rill,y;dy,l01rp;ve0Ixy;ce,y;d,fe,int0l1Ev0U;a9e7i6o4ude;mantic,o16sy,u4;gh,nd;ch,pe,tzy;a4d,mo0A;dy,l;gg5ndom,p4re,w;id;ed;ai2i4;ck,et;hoBi1ClAo9r6u4;ny,r4;e,p3;egna2ic5o4;fouSud;ey,k0;liXor;ain,easa2;ny;dd,i0ld,ranL;aive,e6i5o4;b3isy,rm0Vsy;ce,mb3;a4w;r,t;ad,e6ild,o5u4;nda0Yte;ist,o1;a5ek,l4;low;s0ty;a8ewd,i7o4ucky;f0Gn5o12u4ve0w0Wy0K;d,sy;e0g;ke0tt3ve0;me,r4te;ge;e5i4;nd;en;ol0ui1B;cy,ll,n4;secu7t4;e4ima5;llege2rmedia4;te;re;aBe8i7o6u4;ge,m4ng1E;b3id;me0t;gh,l0;a4fVsita2;dy,v4;en0y;nd15ppy,r4;d,sh;aEenDhBiAl9oofy,r4;a7e6is0o4ue12;o4ss;vy;at,en,y;nd,y;ad,ib,ooE;a2d1;a4o4;st0;t3uiS;u1y;aDeeb3i9lat,o7r6u4;ll,n4r0S;!ny;aDesh,iend0;a4rmEul;my;erce5nan4;ciB;! ;le;ke,n08r,st,ul4;ty;a7erie,sse5v4xtre0G;il;nti4;al;r5s4;tern,y;ly,th0;aCe9i6ru5u4;ll,mb;nk;r5vi4;ne;e,ty;a4ep,nB;d4f,r;!ly;ppVrk;aDhAl8o6r5u4;dd0r0te;isp,uel;ar4ld,mmon,st0ward0zy;se;e4ou1;ar,vO;e4il0;ap,e4;sy;gey,lm,ri4;ng;aJiHlEoCr6u4;r0sy;ly;a8i5o4;ad,wn;g5llia2;nt;ht;sh,ve;ld,un4;cy;a5o4ue;nd,o1;ck,nd;g,tt4;er;d,ld,w1;dy;bsu7ng6we4;so4;me;ry;rd",
+  "Comparable": "true¦0:39;1:3O;2:3D;3:2D;a3Sb3Ac2Zd2Qe2Jf27g1Vh1Li1Fj1Ek1Bl14m0Yn0To0Sp0Jqu0Hr08sJtEuDvBw5y4za0R;el11ou38;a8e6hi1Hi4ry;ck0Dde,l4n1ry,se;d,y;a4i3R;k,ry;nti32ry;a4erda2ulgar;gue,in,st;g0pcomi2Z;a7en2Shi6i5ough,r4;anqu28en1ue;dy,g34me0ny,r03;ck,rs24;ll,me,rt,wd3G;aRcarQePhNiMkin0BlImGoEpDt7u5w4;eet,ift;b4dd0Vperfi1Wrre24;sta22t3;a8e7iff,r5u4;pUr1;a4ict,o2N;ig2Un0N;a1ep,rn;le,rk,te0;e1Oi2Uright0;ci1Vft,l4on,re;emn,id;a4el0;ll,rt;e6i4y;g2Lm4;!y;ek,nd2R;ck,l0mp3;a4iRrill,y;dy,l01rp;ve0Ixy;ce,y;d,fe,int0l1Ev0U;a9e7i6o4ude;mantic,o16sy,u4;gh,nd;ch,pe,tzy;a4d,mo0A;dy,l;gg5ndom,p4re,w;id;ed;ai2i4;ck,et;hoBi1BlAo9r6u4;ny,r4;e,p3;egna2ic5o4;fouSud;ey,k0;liXor;ain,easa2;ny;dd,i0ld,ranL;aive,e6i5o4;b3isy,rm0Vsy;ce,mb3;a4w;r,t;ad,e6ild,o5u4;nda0Xte;ist,o1;a5ek,l4;low;s0ty;a8ewd,i7o4ucky;f0Gn5o11u4ve0w0Vy0K;d,sy;e0g;ke0tt3ve0;me,r4te;ge;e5i4;nd;en;ol0ui19;cy,ll,n4;secu7t4;e4ima5;llege2rmedia4;te;re;aBe8i7o6u4;ge,m4ng1C;b3id;me0t;gh,l0;a4fVsita2;dy,v4;en0y;nd13ppy,r4;d,sh;aEenDhBiAl9oofy,r4;a7e6is0o4ue10;o4ss;vy;at,en,y;nd,y;ad,ib,ooE;a2d1;a4o4;st0;t3uiR;u1y;aDeeb3i9lat,o7r6u4;ll,n4r0Q;!ny;aDesh,iend0;a4rmE;my;erce5nan4;ciB;! ;le;r,ul4;ty;a7erie,sse5v4xtre0E;il;nti4;al;r5s4;tern,y;ly,th0;aBe8i5ru4umb;nk;r5vi4;ne;e,ty;a4ep,nB;d4f,r;!ly;ppUrk;aCh9l8o6r5u4;dd0r0te;isp,uel;ar4ld,mmon,st0ward0zy;se;evOou1;e4il0;ap,e4;sy;gey,ri4;ng;aJiHlEoCr6u4;r0sy;ly;a8i5o4;ad,wn;g5llia2;nt;ht;sh,ve;ld,un4;cy;a5o4ue;nd,o1;ck,nd;g,tt4;er;d,ld,w1;dy;bsu7ng6we4;so4;me;ry;rd",
   "Infinitive": "true¦0:8V;1:8G;2:9D;3:80;4:7N;5:91;6:9H;7:99;8:82;9:9G;A:92;B:8X;C:7K;D:7U;E:7Q;F:87;G:7V;H:7H;a82b7Jc6Sd5Le4Gf45g41h3Wi3Cj39k37l2Ym2Rnou3Vo2Lp25qu24r19s0AtYuTvRwI;aOeNiLors4rI;eJiI;ng,te;ak,st3;d4e7HthI;draw,er;a2d,ep;i2ke,nIrn;d0t;aIie;li9Eni9ry;nIplift;cov0dJear7JlIplug,tie,ve85;ea9o3K;erIo;cut,go,sta9Gval96whelm;aRePhMoLrI;aIemb3;ffi3Fmp3nsI;aDpi8;pp3ugh4;aJiJrIwaE;eat4i2;nk;aIm92;ch,se;ck3ilor,keImp0r7N;! paE;a0Fc0Ee0Ch08i06l04mugg3n03o01pYquXtQuKwI;all71eeIim;p,t4;bKccumb,ffJggeBmm93p2FrI;mouFvi2;er,i5;li80mJsiIveE;de,st;erAit;aMe7SiKrI;ang3eIi2;ng20w;fIng;f4le;gg0rI;t3ve;a3Ui9;a4DeJit,l6KoI;il,of;ak,nd;lIot6Nw;icCve;eak,i0L;aIi6;m,y;ft,nIt;g,k;aKi5LoJriIun;nk,v5X;ot,rt4;ke,rp4tt0;eIll,nd,que7Nv0w;!k,m;avenAul81;dd4tis1Ay;a0BeJip4oI;am,ut;a07b05c03d02fZgroup,heaYiXlVmUnTpRq36sNtLup,vI;amp,eJiIo2V;sCve;l,rt;i8rI;ie2ofC;eFiKtIurfa5;o1XrI;aHiDuctu8;de,gn,st;el,hra1lIreseF;a5e69;d0ew,o05;aHe31o2;a7eFiIoad,y;e2nq3Mve;mbur1nf2U;r1t;inJleDocus,re9uI;el,rbi9;an5e;aDu5;ei2k7Jla3OoIyc3;gni2Wnci3up,v0;oot,uI;ff;ct,liI;se,ze;tt3vi9;a9enGit,o7;aVerTinpoiFlumm0VoSrKuI;b3Oke,ni9rIt;poEs6W;eMoI;cKd,fe4BhibCnoJpo1sp0tru6vI;e,i6o4W;un5;la39u8;aGclu6dJf0occupy,sup0FvI;a5LeF;etermi47iD;aGrtr4Vsse44;cei2fo3Hi9mea7plex,sIva6;eve8iB;mp0rItrol,ve,y;a5St5O;bMccuLpKutJverIwe;lap,s19tu6Au1;gr4Pnu22pa5;era7i3Ypo1;py,r;ser2taH;aMe09iJoIultiply;leBu64;micJnIspla5;ce,g3us;!k;iIke,naA;m,ntaH;aPeLiIo1u39;e,ke,ng0quIv4;eIi66;fy;aKnIss4;d,gI;th4;rn,ve;ng26u1E;eep,nI;e3Row;oHuI;gg3xtaI;po1;gno8mWnI;cUdSfRgeBhQitia7ju8q14sOtKun5TvI;eIo0T;nt,st;erJimi5QoxiQrI;odu5u6;aDn,prIru5Q;et;iBpi8tIu8;il,ruD;abCibC;eBo2Bu1;iIulA;ca7;i7lu6;b5Cmer1pI;aEer47ly,oJrI;e2Ris5No2;rt,se,veI;ri9;aLeKiIoiBuE;de,jaInd0;ck;ar,iQ;mp0ng,pp4st4ve;ath0et,i2le1UoKrI;aIow;b,pp3ze;!ve4S;ast4er3Li58lRorMrJuI;lf3Tr3P;eJiIolic;ght4;e0Lsh4;b3DeKfeCgIs4B;eIi2;!t;clo1go,sIwa4J;had2Y;ee,i2Q;a0KdCl0Im0CnUquip,rTsStGvQxI;cNeEha3iMpJtI;ing0Uol;eJi8lIo1unA;aHoC;ct,di7;st,t;e2MlI;a0Xu6;alua7oI;ke,l2;chew,pou1tab13;a1u4G;aYcVdTfSgQhan5joy,lPqOrNsuMtKvI;e0VisI;aAi4L;er,i5rI;aHenGuB;e,re;iGol;ui8;arAiB;aAeIra2ulf;nd0;or5;ang0oIu8;r1w;lo1ou0CrJuI;mb0;oaGy3Z;b3ct;bKerApI;hasiIow0;ze;a0Tody,rI;a5oiI;d0l;ap1eDuI;ci3Ode;rIt;ma0Nn;a0Ne02iKo,rIwind3;aw,edAoI;wn;agno1e,ff0g,mi27sLvI;eIulA;rIst;ge,t;ab3bUcPlodAmant3pNru3GsMtI;iIoEu2W;lJngI;ui9;!l;ol2ua6;eIla5o1ro2;n1r1;a2Ne2WlKoIu0K;uIv0;raA;aIo1;im;a38ur1;af4bZcTduDep4fRliQmNnLpKra1TtaGvI;eIol2;lop;aEiDoE;oIy;te,un5;eJoI;li9;an;mCv0;a5i06oIraud,y;rm;ei2iMoKrI;ee,yI;!pt;de,mIup3;missi2Upo1;de,ma7ph0;aJrief,uI;g,nk;rk;mp4rk4uF;a06ea1h04i03l02oJrIurta17;a2ea7ipp3;ales5eZhabCinci6llYmWnIrro6;cTdQfNju8no7qu0sLtKvI;eIin5;ne,rA;aHin25ribu7;er2iIoli27pi8titu7ult;d0st;eJiIroFu1;de,gu8rm;ss;eJoI;ne;mn,n1;eIlu6ur;al,i2;buBe,men5pI;e7i3ly;eDi6u6;r5xiB;ean1iS;rcumveFte;eIoo1;ri9w;ncIre4t0ulk;el;aXeRi6lPoOrLuI;iJrIy;st,y;ld;aJeastfeMiIoad4;ng;ke;il,l12mba0XrrMth0;eIow;ed;!come,gHha2liLqueaKstJtrIwild0;ay;ow;th;e2tt3;in;bysCckfi8ff3tI;he;it;b17c0Vd0Mff0Kgr0Jl0Fm0Bn05pp01rZsSttPuNvKwaI;it,k4;en;eEoI;id;rt;gIto08;meF;aGeBraD;ct;ch;pi8sJtoI;ni9;aKeIi05u8;mb3rt,ss;le;il;re;g0Ji1ou1rI;anAi2;eaKly,oiFrI;ai1o2;nt;r,se;aMiQnJtI;icipa7;eJoIul;un5y;al;ly1;aJu1;se;lgaIze;ma7;iKlI;eAoIu6;t,w;gn;ee;ix,oI;rd;a01jNmiKoJsoI;rb;pt,rn;niIt;st0;er;ouJuB;st;rn;cLhie2knowledAquiItiva7;es5re;ce;ge;eOomKrJusI;e,tom;ue;moJpI;any,li9;da7;te;pt;andOet,i6oKsI;coKol2;ve;li9rt,uI;nd;sh;de;on",
   "Modal": "true¦c5lets,m4ought3sh1w0;ill,o5;a0o4;ll,nt;! to,a;ight,ust;an,o0;uld",
   "Verb": "true¦born,cannot,gonna,has,keep tabs,msg",
